% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHANGED-pre-msa-tree.R
\name{acc2FA}
\alias{acc2FA}
\title{acc2FA converts protein accession numbers to a fasta format.}
\usage{
acc2FA(accessions, outpath, plan = "sequential")
}
\arguments{
\item{accessions}{Character vector containing protein accession numbers to
generate fasta sequences for.
Function may not work for vectors of length > 10,000}

\item{outpath}{\link{str} Location where fasta file should be written to.}

\item{plan}{Character string specifying the parallel processing strategy to
use with the \code{future} package. Default is "sequential".}
}
\value{
A logical value indicating whether the retrieval and conversion were
successful. Returns \code{TRUE} if successful and \code{FALSE} otherwise.
}
\description{
Resulting fasta file is written to the outpath.
}
\examples{
\dontrun{
acc2FA(accessions = c("ACU53894.1", "APJ14606.1", "ABK37082.1"), 
outpath = "my_proteins.fasta")
Entrez:accessions <- rep("ANY95992.1", 201) |> acc2FA(outpath = "entrez.fa")
EBI:accessions <- c("P12345", "Q9UHC1", 
"O15530", "Q14624", "P0DTD1") |> acc2FA(outpath = "ebi.fa")
}
}
\author{
Samuel Chen, Janani Ravi
}
\keyword{accnum,}
\keyword{fasta}
