% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_leaves.R
\name{add_leaves}
\alias{add_leaves}
\title{Adding Leaves to an alignment file w/ accessions}
\usage{
add_leaves(input_file = "data/alignments/pspa_snf7.aln",
  lin_file = "data/pspa_snf7.txt")
}
\arguments{
\item{input_file}{Character. Path to file. Input tab-delimited file +
alignment file accnum & alignment.
Default is 'pspa_snf7.aln'}

\item{lin_file}{Character. Path to file. Protein file with accession +
number to lineage mapping.
Default is 'pspa.txt'}
}
\description{
Adding Leaves to an alignment file w/ accessions
Genomic Contexts vs Domain Architectures.
}
\details{
The alignment file would need two columns: 1. accession +
number and 2. alignment. The protein homolog accession to lineage mapping +
file should have
}
\note{
Please refer to the source code if you have alternate +
file formats and/or column names.
}
\examples{
\dontrun{
add_leaves("pspa_snf7.aln", "pspa.txt")
}
\author{
Janani Ravi
}
\keyword{accnum,}
\keyword{alignment,}
\keyword{leaves,}
\keyword{lineage,}
\keyword{species}
