% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineage.R
\name{add_lins}
\alias{add_lins}
\title{add_lins}
\usage{
add_lins(
  df,
  acc_col = "AccNum",
  assembly_path,
  lineagelookup_path,
  ipgout_path = NULL,
  plan = "multicore"
)
}
\arguments{
\item{plan}{}
}
\description{
add_lins
}
