% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre-msa-tree.R
\name{add_name}
\alias{add_name}
\title{Title}
\usage{
add_name(
  data,
  accnum_col = "AccNum",
  spec_col = "Species",
  lin_col = "Lineage",
  lin_sep = ">",
  out_col = "Name"
)
}
\arguments{
\item{data}{Data to add name column to}

\item{accnum_col}{Column containing accession numbers}

\item{spec_col}{Column containing species}

\item{lin_col}{Column containing lineage}

\item{lin_sep}{Character separating lineage levels}

\item{out_col}{Column that contains the new 'Name' derived from Species,
Lineage, and AccNum info}
}
\value{
Original data with a 'Name' column
}
\description{
This function adds a new 'Name' column that is comprised of components from
Kingdom, Phylum, Genus, and species, as well as the accession
}
\author{
Samuel Chen, Janani Ravi
}
