% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHANGED-pre-msa-tree.R, R/pre-msa-tree.R
\name{addLeaves2Alignment}
\alias{addLeaves2Alignment}
\title{addLeaves2Alignment}
\usage{
addLeaves2Alignment(
  aln_file = "",
  lin_file = "data/rawdata_tsv/all_semiclean.txt",
  reduced = FALSE
)

addLeaves2Alignment(
  aln_file = "",
  lin_file = "data/rawdata_tsv/all_semiclean.txt",
  reduced = FALSE
)
}
\arguments{
\item{aln_file}{Character. Path to file. Input tab-delimited file +
alignment file accnum & alignment.
Default is 'pspa_snf7.aln'}

\item{lin_file}{Character. Path to file. Protein file with accession +
number to lineage mapping.
Default is 'pspa.txt'}

\item{reduced}{Boolean. If TRUE, a reduced data frame will be generated with
only one sequence per lineage. Default is FALSE.}
}
\description{
Adding Leaves to an alignment file w/ accessions
Genomic Contexts vs Domain Architectures.

Adding Leaves to an alignment file w/ accessions
Genomic Contexts vs Domain Architectures.
}
\details{
The alignment file would need two columns: 1. accession +
number and 2. alignment. The protein homolog accession to lineage mapping +
file should have

The alignment file would need two columns: 1. accession +
number and 2. alignment. The protein homolog accession to lineage mapping +
file should have
}
\note{
Please refer to the source code if you have alternate +
file formats and/or column names.

Please refer to the source code if you have alternate +
file formats and/or column names.
}
\examples{
\dontrun{
addLeaves2Alignment("pspa_snf7.aln", "pspa.txt")
}
\dontrun{
addLeaves2Alignment("pspa_snf7.aln", "pspa.txt")
}
}
\author{
Janani Ravi
}
\keyword{accnum,}
\keyword{alignment,}
\keyword{leaves,}
\keyword{lineage,}
\keyword{species}
