% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acc2lin.R, R/lineage.R
\name{addLineage}
\alias{addLineage}
\title{addLineage}
\usage{
addLineage(
  df,
  acc_col = "AccNum",
  assembly_path,
  lineagelookup_path,
  ipgout_path = NULL,
  plan = "multicore"
)

addLineage(
  df,
  acc_col = "AccNum",
  assembly_path,
  lineagelookup_path,
  ipgout_path = NULL,
  plan = "multicore"
)
}
\arguments{
\item{df}{Dataframe containing accession numbers. The dataframe should
have a column specified by \code{acc_col} that contains these accession numbers.}

\item{acc_col}{Character. The name of the column in \code{df} containing
accession numbers. Default is "AccNum".}

\item{assembly_path}{String. The path to the assembly summary file generated
using the \code{downloadAssemblySummary()} function.}

\item{lineagelookup_path}{String. The path to the lineage lookup file (taxid
to lineage mapping) generated using the \code{create_lineage_lookup()} function.}

\item{ipgout_path}{String. Optional path to save intermediate output files.
Default is NULL.}

\item{plan}{Character. Specifies the execution plan for parallel processing.
Default is "multicore".}
}
\value{
A \code{data.frame} that combines the original \code{df} with the lineage
information.

A dataframe that combines the original dataframe \code{df} with lineage
information retrieved based on the provided accession numbers.
}
\description{
addLineage

addLineage
}
\examples{
\dontrun{
addLineage()
}
\dontrun{
enriched_df <- addLineage(df = my_data,
                           acc_col = "AccNum",
                           assembly_path = "path/to/assembly_summary.txt",
                           lineagelookup_path = "path/to/lineage_lookup.tsv")
}
}
