% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHANGED-pre-msa-tree.R, R/pre-msa-tree.R
\name{addName}
\alias{addName}
\title{addName}
\usage{
addName(
  data,
  accnum_col = "AccNum",
  spec_col = "Species",
  lin_col = "Lineage",
  lin_sep = ">",
  out_col = "Name"
)

addName(
  data,
  accnum_col = "AccNum",
  spec_col = "Species",
  lin_col = "Lineage",
  lin_sep = ">",
  out_col = "Name"
)
}
\arguments{
\item{data}{Data to add name column to}

\item{accnum_col}{Column containing accession numbers}

\item{spec_col}{Column containing species}

\item{lin_col}{Column containing lineage}

\item{lin_sep}{Character separating lineage levels}

\item{out_col}{Column that contains the new 'Name' derived from Species,
Lineage, and AccNum info}
}
\value{
Original data with a 'Name' column

Original data with a 'Name' column
}
\description{
This function adds a new 'Name' column that is comprised of components from
Kingdom, Phylum, Genus, and species, as well as the accession

This function adds a new 'Name' column that is comprised of
components from Kingdom, Phylum, Genus, and species, as well as the accession
}
\examples{
# Example usage of the addName function
data <- data.frame(
  AccNum = c("ACC123", "ACC456"),
  Species = c("Homo sapiens", "Mus musculus"),
  Lineage = c("Eukaryota>Chordata", "Eukaryota>Chordata")
)
enriched_data <- addName(data)
enriched_data
\dontrun{
addName(data_frame)
}
}
\author{
Samuel Chen, Janani Ravi
}
