% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{aggregate_logs}
\alias{aggregate_logs}
\title{Stack rows of job logfiles}
\usage{
aggregate_logs(
  dir_job_results,
  latest_date = (Sys.Date() - 60),
  verbose = FALSE
)
}
\arguments{
\item{dir_job_results}{the base directory containing all job_results folders as subdirs}
}
\value{
list of log table and a character vector of logfile paths that failed to read
}
\description{
Stack rows of job logfiles
}
