% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHANGED-pre-msa-tree.R, R/pre-msa-tree.R
\name{alignFasta}
\alias{alignFasta}
\title{alignFasta}
\usage{
alignFasta(fasta_file, tool = "Muscle", outpath = NULL)

alignFasta(fasta_file, tool = "Muscle", outpath = NULL)
}
\arguments{
\item{fasta_file}{Path to the FASTA file to be aligned}

\item{tool}{Type of alignment tool to use. One of three options: "Muscle", "ClustalO", or "ClustalW"}

\item{outpath}{Path to write the resulting alignment to as a FASTA file. If NULL, no file is written}
}
\value{
aligned fasta sequence as a MsaAAMultipleAlignment object

aligned fasta sequence as a MsaAAMultipleAlignment object
}
\description{
Perform a Multiple Sequence Alignment on a FASTA file.

Perform a Multiple Sequence Alignment on a FASTA file.
}
\author{
Samuel Chen, Janani Ravi
}
