% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_job_queue.R
\name{assign_job_queue}
\alias{assign_job_queue}
\title{Decision function to assign job queue}
\usage{
assign_job_queue(t_sec_estimate, t_cutoff = 21600)
}
\arguments{
\item{t_sec_estimate}{estimated number of seconds a job will process
(from advanced_opts2est_walltime())}

\item{t_long}{threshold value that defines the lower bound for assigning a
job to the "long queue"}
}
\value{
a string of "short" or "long"

example:
advanced_opts2est_walltime(c("homology_search", "domain_architecture"), 3) |>
assign_job_queue()
}
\description{
Decision function to assign job queue
}
