% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_job_queue.R
\name{assignJobQueue}
\alias{assignJobQueue}
\title{assignJobQueue}
\usage{
assignJobQueue(t_sec_estimate, t_cutoff = 21600)
}
\arguments{
\item{t_sec_estimate}{estimated number of seconds a job will process
(from calculateEstimatedWallTimeFromOpts    ())}

\item{t_long}{threshold value that defines the lower bound for assigning a
job to the "long queue"}
}
\value{
a string of "short" or "long"

example:
calculateEstimatedWallTimeFromOpts	(c("homology_search",
"domain_architecture"), 3) |>
assignJobQueue()
}
\description{
Decision function to assign job queue
}
