% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{calc_log_process_stat}
\alias{calc_log_process_stat}
\title{calc statistic for processes in MolEvolvR logfiles}
\usage{
calc_log_process_stat(df_log, columns, f, ..., columns_group_by = character(0))
}
\arguments{
\item{df_log}{\code{df_log} element from the return list of \code{aggregate_logs()}}
}
\value{
if group_by is empty: a single value; else: tibble from summarise
}
\description{
calc statistic for processes in MolEvolvR logfiles
}
