% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_job_queue.R
\name{calculateEstimatedWallTimeFromOptions}
\alias{calculateEstimatedWallTimeFromOptions}
\title{Given MolEvolvR advanced options and number of inputs,
calculate the total estimated walltime for the job}
\usage{
calculateEstimatedWallTimeFromOptions(
  advanced_opts,
  n_inputs = 1L,
  n_hits = NULL,
  verbose = FALSE
)
}
\arguments{
\item{advanced_opts}{character vector of MolEvolvR advanced options
(see mapOption2Process for the options)}

\item{n_inputs}{total number of input proteins}
}
\value{
total estimated number of seconds a job will process (walltime)

example: calculateEstimatedWallTimeFromOptions(c("homology_search",
"domain_architecture"),
n_inputs = 3, n_hits = 50L)
}
\description{
Given MolEvolvR advanced options and number of inputs,
calculate the total estimated walltime for the job
}
