% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanup.R
\name{clean_string}
\alias{clean_string}
\title{Clean String}
\usage{
clean_string(string)
}
\arguments{
\item{string}{}
}
\value{
\link{string} string with only alphanumerics, "_", "+", and "."
}
\description{
Keep only alphanumerics, "\emph{", "+", and "." in strings
and substitute spaces with "}". Used in MolEvolvR codebase to
cleanup domain architecture values
}
\examples{
\dontrun{
clean_string()
}

}
