% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanup.R
\name{cleanGeneDescription}
\alias{cleanGeneDescription}
\title{Cleanup GeneDesc}
\usage{
cleanGeneDescription(prot, column)
}
\arguments{
\item{prot}{A data frame containing the gene descriptions.}

\item{column}{The name of the column from which gene descriptions are pulled
for cleanup.}
}
\value{
Return trailing period that occurs in GeneDesc column
}
\description{
Cleanup GeneDesc
}
\examples{
\dontrun{
cleanGeneDescription()
}
}
