% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanup.R
\name{cleanLineage}
\alias{cleanLineage}
\title{Cleanup Lineage}
\usage{
cleanLineage(prot, lins_rename)
}
\arguments{
\item{prot}{A data frame containing a 'Lineage' column that needs to be
cleaned up.}

\item{lins_rename}{A data frame with two columns: 'old' containing terms
to be replaced and 'new' containing the corresponding replacement terms.}
}
\value{
The original data frame with the 'Lineage' column updated based on
the provided replacements.
}
\description{
Cleanup Lineage
}
\examples{
\dontrun{
cleanLineage()
}
}
