% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanup.R
\name{cleanSpecies}
\alias{cleanSpecies}
\title{Cleanup Species}
\usage{
cleanSpecies(prot, removeEmptyRows = FALSE)
}
\arguments{
\item{prot}{A data frame that contains columns 'Species'.}

\item{removeEmptyRows}{Boolean. If TRUE, rows with empty/unnecessary values in 'Species' are removed.
Default is false.}
}
\value{
The original data frame with Species cleaned.
}
\description{
Cleanup Species

Cleans up the species column of a data frame by removing certain characters and rows.

This function removes unneccessary characters from the 'Species' column.
A cleaned up version of the data table is returned.
}
\examples{
\dontrun{
cleanSpecies(prot, TRUE)
}
}
