% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanup.R
\name{cleanString}
\alias{cleanString}
\title{Clean String}
\usage{
cleanString(string)
}
\arguments{
\item{string}{}
}
\value{
\link[rlang:string]{rlang::string} string with only alphanumerics, "_", "+", and "."
}
\description{
Keep only alphanumerics, "\emph{", "+", and "." in strings
and substitute spaces with "}". Used in MolEvolvR codebase to
cleanup domain architecture values
}
\examples{
\dontrun{
cleanString()
}

}
