% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanup.R
\name{cleanup_clust}
\alias{cleanup_clust}
\title{Cleanup Clust}
\usage{
cleanup_clust(
  prot,
  domains_rename,
  domains_keep,
  repeat2s = TRUE,
  remove_tails = FALSE,
  remove_empty = FALSE
)
}
\arguments{
\item{prot}{A data frame that must contain columns Query and ClustName.}

\item{domains_rename}{A data frame containing the domain names to be replaced in a column 'old' and the corresponding replacement values in a column 'new'.}

\item{domains_keep}{A data frame containing the domain names to be retained.}

\item{repeat2s}{Boolean. If TRUE, repeated domains in 'ClustName' are condensed. Default is TRUE.}

\item{remove_tails}{Boolean. If TRUE, 'ClustName' will be filtered based on domains to keep/remove. Default is FALSE.}

\item{remove_empty}{Boolean. If TRUE, rows with empty/unnecessary values in 'ClustName' are removed. Default is FALSE.}
}
\value{
Cleaned up data frame
}
\description{
Cleanup cluster file

Cleans a cluster file by removing rows that do not contain the query in the cluster.

This function removes irrelevant rows which do not contain the query protein within the ClustName column.
The return value is the cleaned up data frame.
}
\examples{
\dontrun{
cleanup_clust(prot, TRUE, FALSE, domains_keep, domains_rename)
}
}
