% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanup.R
\name{cleanup_GeneDesc}
\alias{cleanup_GeneDesc}
\title{Cleanup GeneDesc}
\usage{
cleanup_GeneDesc(prot, column)
}
\arguments{
\item{column}{}
}
\value{
Return trailing period that occurs in GeneDesc column
}
\description{
Cleanup GeneDesc
}
\examples{
\dontrun{
cleanup_GeneDesc()
}
}
