% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanup.R
\name{cleanup_lineage}
\alias{cleanup_lineage}
\title{Cleanup Lineage}
\usage{
cleanup_lineage(prot, lins_rename)
}
\arguments{
\item{lins_rename}{}
}
\value{
Describe return, in detail
}
\description{
Cleanup Lineage
}
\examples{
\dontrun{
cleanup_lineage()
}
}
