% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanup.R
\name{cleanup_species}
\alias{cleanup_species}
\title{Cleanup Species}
\usage{
cleanup_species(prot, remove_empty = FALSE)
}
\arguments{
\item{prot}{A data frame that contains columns 'Species'.}

\item{remove_empty}{Boolean. If TRUE, rows with empty/unnecessary values in 'Species' are removed.
Default is false.}
}
\value{
Describe return, in detail
}
\description{
Cleanup Species

Cleans up the species column of a data frame by removing certain characters and rows.

This function removes unneccessary characters from the 'Species' column.
A cleaned up version of the data table is returned.
}
\examples{
\dontrun{
cleanup_species(prot, TRUE)
}
}
