% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_files.R
\name{combine_files}
\alias{combine_files}
\title{Download the combined assembly summaries of genbank and refseq}
\usage{
combine_files(
  inpath = c("../molevol_data/project_data/phage_defense/"),
  pattern = "*full_analysis.tsv",
  delim = "\\t",
  skip = 0,
  col_names = T
)
}
\arguments{
\item{inpath}{String of 'master' path where the files reside (recursive=T)}

\item{pattern}{Character vector containing search pattern for files}

\item{col_names}{Takes logical T/F arguments OR column names vector;
usage similar to col_names parameter in \code{readr::read_delim}}
}
\description{
Download the combined assembly summaries of genbank and refseq
}
\author{
Janani Ravi
}
