% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_analysis.R
\name{combine_ipr}
\alias{combine_ipr}
\title{Combining clean ipr files}
\usage{
combine_ipr(inpath, ret = FALSE)
}
\arguments{
\item{inpath}{Character. The path to the directory containing the
\code{.iprscan_cln.tsv} files to be combined.}

\item{ret}{Logical. If TRUE, the function will return the combined data frame.
Default is FALSE, meaning it will only write the file and not return the data.}
}
\value{
If \code{ret} is TRUE, a data frame containing the combined data from all
input files. If \code{ret} is FALSE, the function writes the combined data to a
TSV file named \code{ipr_combined.tsv} in the specified directory and returns NULL.
}
\description{
Combining clean ipr files
}
\examples{
\dontrun{
combined_ipr_data <- combine_ipr("path/to/ipr/files", ret = TRUE)
}
}
