% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanup.R
\name{condenseRepeatedDomains}
\alias{condenseRepeatedDomains}
\title{condenseRepeatedDomains}
\usage{
condenseRepeatedDomains(prot, by_column = "DomArch", excluded_prots = c())
}
\arguments{
\item{prot}{A data frame containing 'DomArch', 'GenContext', 'ClustName' columns.}

\item{by_column}{Column in which repeats are condensed to domain+domain -> domain(s).}

\item{excluded_prots}{Vector of strings that condenseRepeatedDomains should not reduce to (s). Defaults to c()}
}
\value{
A data frame with condensed repeated domains in the specified column.
}
\description{
Condense repeated domains

Condenses repeated domains in the specified column.

This function identifies repeated domains and condenses them to (s).
?? Certain domains can be removed according to an additional data frame.
The original data frame is returned with the corresponding cleaned up column.
}
\examples{
\dontrun{
condenseRepeatedDomains(prot, "DomArch")
}
}
