% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHANGED-pre-msa-tree.R
\name{convert2TitleCase}
\alias{convert2TitleCase}
\alias{totitle,}
\alias{to_title}
\title{Changing case to 'Title Case'}
\usage{
convert2TitleCase(text, delimitter)
}
\arguments{
\item{x}{Character vector.}

\item{y}{Delimitter. Default is space (" ").}
}
\description{
Translate string to Title Case w/ delimitter.
}
\seealso{
chartr, toupper, and tolower.
}
\author{
Andrie, Janani Ravi
}
