% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{convert_fa2tre}
\alias{convert_fa2tre}
\title{convert_fa2tre}
\usage{
convert_fa2tre(
  fa_path = here("data/alns/pspa_snf7.fa"),
  tre_path = here("data/alns/pspa_snf7.tre"),
  fasttree_path = here("src/FastTree")
)
}
\arguments{
\item{fa_path}{Path to the input FASTA alignment file (.fa). Default is the
path to "data/alns/pspa_snf7.fa".}

\item{tre_path}{Path to the output file where the generated tree (.tre) will
be saved. Default is the path to "data/alns/pspa_snf7.tre".}

\item{fasttree_path}{Path to the FastTree executable, which is used to
generate the phylogenetic tree. Default is "src/FastTree".}
}
\value{
No return value. The function generates a tree file (.tre) from the
input FASTA file.
}
\description{
convert_fa2tre
}
\examples{
\dontrun{
convert_fa2tre(here("data/alns/pspa_snf7.fa"), 
                here("data/alns/pspa_snf7.tre"), 
                here("src/FastTree")
}
}
