% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{convertAlignment2Trees}
\alias{convertAlignment2Trees}
\title{convertAlignment2Trees}
\usage{
convertAlignment2Trees(aln_path = here("data/alns/"))
}
\arguments{
\item{aln_path}{}
}
\description{
Generate Trees for ALL fasta files in "data/alns"
}
