% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{convertFA2Tree}
\alias{convertFA2Tree}
\title{convertFA2Tree}
\usage{
convertFA2Tree(
  fa_path = here("data/alns/pspa_snf7.fa"),
  tre_path = here("data/alns/pspa_snf7.tre"),
  fasttree_path = here("src/FastTree")
)
}
\arguments{
\item{fasttree_path}{}
}
\description{
convertFA2Tree
}
