% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{countByColumn}
\alias{countByColumn}
\title{Count By Column}
\usage{
countByColumn(prot = prot, column = "DomArch", min.freq = 1)
}
\arguments{
\item{min.freq}{}
}
\value{
Describe return, in detail
}
\description{
Count By Column
}
\examples{
\dontrun{
countByColumn()
}
}
