% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotme.R
\name{create_all_col_params}
\alias{create_all_col_params}
\title{create_all_col_params}
\usage{
create_all_col_params(count_data, fill_by_n, sort_by_n)
}
\arguments{
\item{sort_by_n}{}
}
\description{
create_all_col_params
}
