% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{createDomarchTable}
\alias{createDomarchTable}
\title{Creates a formatted table for domain architectures.}
\usage{
createDomarchTable(job_dir)
}
\arguments{
\item{job_dir}{Character. The directory containing the job files.}
}
\value{
A formatted table object.
}
\description{
This function generates a formatted table summarizing domain architectures
from a given job directory. It ensures that the necessary preprocessing
steps (filtering by genome and domains) are performed before creating the table.
}
\examples{
\dontrun{
createDomarchTable("path/to/job_dir")
}
}
