% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks_gencontext.R
\name{createGenomicContextNetwork}
\alias{createGenomicContextNetwork}
\title{Genomic Context Directed Network}
\usage{
createGenomicContextNetwork(
  prot,
  domains_of_interest,
  column = "GenContext",
  cutoff = 40,
  layout = "grid",
  directed = TRUE
)
}
\arguments{
\item{prot}{A data frame that contains the column 'GenContext'.}

\item{domains_of_interest}{Character vector of domains of interest.}

\item{column}{Name of column containing Genomic Context from which nodes and
edges are generated.}

\item{cutoff}{Integer. Only use GenContexts that occur at or above the cutoff
percentage for total count}

\item{layout}{Character. Layout type to be used for the network. Options are:
\itemize{\item "grid" \item "circle" \item "random" \item "auto" \item "nice"}}

\item{directed}{Is the network directed?}
}
\value{
A plot of the genomic context network.
}
\description{
This function creates a Genomic Context network from the 'GenContext' column.

A network of Genomic Context is returned.
}
\examples{
\dontrun{
gc_directed_network(pspa, column = "GenContext", cutoff = 55)
}
}
