% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa2domain.R
\name{createIPRScanDomainTable}
\alias{createIPRScanDomainTable}
\title{For a given accession number, get the domain sequences using a interproscan
output table & the original FASTA file}
\usage{
createIPRScanDomainTable(
  accnum,
  fasta,
  df_iprscan,
  analysis = c("Pfam", "Gene3D")
)
}
\arguments{
\item{accnum}{\link{chr} a \emph{single} accession number from the original fasta (fasta param)
which will be used to search for its sequence's domains (df_iprscan param)}

\item{fasta}{\link{AAStringSet} original fasta file which was fed into interproscan}

\item{df_iprscan}{\link{tbl_df} the output TSV of interproscan, read as a tibble with
readIPRScanTSV()}

\item{analysis}{\link{chr} the domain databases to extract sequences from}
}
\value{
\link{tbl_df} table with each domain sequence and a new identifier column
}
\description{
For a given accession number, get the domain sequences using a interproscan
output table & the original FASTA file
}
\examples{
\dontrun{
path_molevol_scripts <- file.path(Sys.getenv("DEV", unset = "/data/molevolvr_transfer/molevolvr_dev"), "molevol_scripts")
setwd(path_molevol_scripts)
source("R/fa2domain.R")
fasta <- Biostrings::readAAStringSet("./tests/example_protein.fa")
df_iprscan <- readIPRScanTSV("./tests/example_iprscan_valid.tsv")
accnum <- df_iprscan$AccNum[1]
df_iprscan_domains <- createIPRScanDomainTable(accnum, fasta, df_iprscan)
}

}
