% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msa.R
\name{createMSA_Kalign}
\alias{createMSA_Kalign}
\title{Function to generate MSA using kalign}
\usage{
createMSA_Kalign(fa_file = "", outfile = "")
}
\arguments{
\item{fa_file}{Character. The path to the input FASTA file containing protein
sequences.}

\item{outfile}{Character. The path to the output file where the alignment
will be saved.}
}
\value{
A list containing the alignment object and the output file path.
}
\description{
Function to generate MSA using kalign
}
\examples{
\dontrun{
createMSA_Kalign(fa_file = "path/to/sequences.fasta", 
                 outfile = "path/to/alignment.txt")
}
}
