% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{createWordCloudElement}
\alias{createWordCloudElement}
\title{Wordclouds for the predominant domains, domain architectures}
\usage{
createWordCloudElement(
  query_data = "prot",
  colname = "DomArch",
  cutoff = 70,
  UsingRowsCutoff = FALSE
)
}
\arguments{
\item{query_data}{Data frame of protein homologs with the usual 11 columns +
additional word columns (0/1 format). Default is "prot".}

\item{colname}{Character. The name of the column in \code{query_data} to generate
the word cloud from. Default is "DomArch".}

\item{cutoff}{Numeric. The cutoff value for filtering elements based on their
frequency. Default is 70.}

\item{UsingRowsCutoff}{Logical. Whether to use a row-based cutoff instead of
a frequency cutoff. Default is FALSE.}
}
\value{
A word cloud plot showing the frequency of elements from the selected
column.
}
\description{
Wordclouds for the predominant domains (from DAs) and DAs (from GC)
}
\details{
For "da2doms" you would need the file DA.doms.wc as well as the
column query_data$DomArch.norep

For "gc2da", you would need the file GC.DA.wc as well as the column
query_data$GenContext.norep
}
\note{
Please refer to the source code if you have alternate file formats and/or
column names.
}
\examples{
\dontrun{
createWordCloudElement(prot, "da2doms", 10)
}
}
\author{
Janani Ravi
}
\keyword{Architectures,}
\keyword{Domain}
\keyword{Domains,}
\keyword{GenomicContexts}
