% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecate.R
\name{deprecate}
\alias{deprecate}
\alias{sink.reset}
\alias{add_lins}
\alias{acc2lin}
\alias{efetch_ipg}
\alias{ipg2lin}
\title{These functions will be deprecated. Please use other functions instead.}
\usage{
sink.reset()

add_lins(
  df,
  acc_col = "AccNum",
  assembly_path,
  lineagelookup_path,
  ipgout_path = NULL,
  plan = "multicore"
)

acc2lin(
  accessions,
  assembly_path,
  lineagelookup_path,
  ipgout_path = NULL,
  plan = "multicore"
)

efetch_ipg(accessions, out_path, plan = "multicore")

ipg2lin(
  accessions,
  ipg_file,
  refseq_assembly_path,
  genbank_assembly_path,
  lineagelookup_path
)
}
\description{
These functions will be deprecated. Please use other functions instead.
}
