% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineage.R
\name{DownloadAssemblySummary}
\alias{DownloadAssemblySummary}
\title{Download the combined assembly summaries of genbank and refseq}
\usage{
DownloadAssemblySummary(
  outpath,
  keep = c("assembly_accession", "taxid", "species_taxid", "organism_name")
)
}
\arguments{
\item{outpath}{String of path where the assembly summary file should be written}

\item{keep}{Character vector containing which columns should be retained and downloaded}
}
\description{
Download the combined assembly summaries of genbank and refseq
}
\author{
Samuel Chen, Janani Ravi
}
