% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acc2lin.R
\name{efetch_ipg}
\alias{efetch_ipg}
\title{efetch_ipg}
\usage{
efetch_ipg(accnums, out_path, plan = "sequential")
}
\arguments{
\item{accnums}{Character vector containing the accession numbers to query on
the ipg database}

\item{out_path}{Path to write the efetch results to}

\item{plan}{}
}
\value{
Describe return, in detail
}
\description{
Perform efetch on the ipg database and write the results to out_path
}
\examples{
\dontrun{
efetch_ipg()
}
}
\author{
Samuel Chen, Janani Ravi
}
