% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{elements2Words}
\alias{elements2Words}
\title{Elements 2 Words}
\usage{
elements2Words(prot, column = "DomArch", conversion_type = "da2doms")
}
\arguments{
\item{prot}{\link{dataframe}}

\item{column}{\link{string} column name}

\item{conversion_type}{\link{string} type of conversion: 'da2doms': domain architectures to
domains. 'gc2da' genomic context to domain architectures}
}
\value{
\link{string} with words delimited by spaces
}
\description{
Break string ELEMENTS into WORDS for domain architecture (DA) and genomic
context (GC)
}
\examples{
\dontrun{
tibble::tibble(DomArch = c("aaa+bbb", "a+b", "b+c", "b-c")) |> elements2Words()
}

}
