% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanup.R
\name{ensureUniqAccNum}
\alias{ensureUniqAccNum}
\title{make accnums unique}
\usage{
ensureUniqAccNum(accnums)
}
\arguments{
\item{accnums}{\link[rlang:vector-construction]{rlang::chr} a vector of accession numbers}
}
\value{
\link[rlang:vector-construction]{rlang::chr} a vector of adjusted, unique accession numbers
}
\description{
Append an index of occurence suffix to each accession number (or any
character vector) making them unique
}
\examples{
\dontrun{
c("xxx", "xxx", "xxx", "yyy", "yyy") |>
    ensureUniqAccNum()
}
}
