% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa2domain.R
\name{exec_interproscan}
\alias{exec_interproscan}
\title{exec_interproscan}
\usage{
exec_interproscan(filepath_fasta, filepath_out, appl = c("Pfam", "Gene3D"))
}
\arguments{
\item{appl}{}
}
\description{
exec_interproscan
}
