% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanup.R
\name{extractAccNum}
\alias{extractAccNum}
\title{extractAccNum}
\usage{
extractAccNum(string)
}
\arguments{
\item{string}{}
}
\value{
Describe return, in detail
}
\description{
extractAccNum
}
\examples{
\dontrun{
extractAccNum()
}
}
