% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{filterByDomains}
\alias{filterByDomains}
\title{Filter by Domains}
\usage{
filterByDomains(
  prot,
  column = "DomArch",
  doms_keep = c(),
  doms_remove = c(),
  ignore.case = FALSE
)
}
\arguments{
\item{prot}{Dataframe to filter}

\item{column}{Column to search for domains in (DomArch column)}

\item{doms_keep}{Vector of domains that must be identified within column in order for
observation to be kept}

\item{doms_remove}{Vector of domains that, if found within an observation, will be removed}

\item{ignore.case}{Should the matching be non case sensitive}
}
\value{
Filtered data frame
}
\description{
filterByDomains filters a data frame by identifying exact domain matches
and either keeping or removing rows with the identified domain
}
\note{
There is no need to make the domains 'regex safe', that will be handled by this function
}
\examples{
\dontrun{
filterByDomains()
}
}
\author{
Samuel Chen, Janani Ravi
}
