% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filterByGenome}
\alias{filterByGenome}
\title{Filter by Genome}
\usage{
filterByGenome(job_dir)
}
\arguments{
\item{job_dir}{The directory containing the job files.}
}
\value{
None. Writes the filtered data to a file.
}
\description{
Filters the data by genome and writes the filtered data to a file.
}
