% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{find_paralogs}
\alias{find_paralogs}
\title{Find Paralogs}
\usage{
find_paralogs(prot)
}
\arguments{
\item{prot}{A data frame filtered by a Query, containing columns Species and Lineage}
}
\description{
Creates a data frame of paralogs.
}
\details{
This function returns a dataframe containing paralogs and the counts.
}
\note{
Please refer to the source code if you have alternate file formats and/or
column names.
}
