% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{findParalogs}
\alias{findParalogs}
\title{findParalogs}
\usage{
findParalogs(prot)
}
\arguments{
\item{prot}{A data frame filtered by a Query, containing columns Species and Lineage}
}
\value{
returns a dataframe containing paralogs and the counts.
}
\description{
Creates a data frame of paralogs.
}
\note{
Please refer to the source code if you have alternate file formats and/or
column names.
}
\examples{
\dontrun{
findParalogs(pspa)
}
}
