% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_status_emails.R
\name{format_job_args}
\alias{format_job_args}
\title{Format job arguments into html-formatted key/value pairs, for including
in an email}
\usage{
format_job_args(job_args)
}
\arguments{
\item{job_args}{a list of job arguments, e.g. as read from the job_args.yml file}
}
\value{
a list of HTML-formatted key/value pairs

example:
format_job_args("/data/scratch/janani/molevolvr_out/Ba5sV1_full")
}
\description{
Format job arguments into html-formatted key/value pairs, for including
in an email
}
