% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_status_emails.R
\name{formatJobArgumentsHTML}
\alias{formatJobArgumentsHTML}
\title{Format job arguments into html-formatted key/value pairs, for including
in an email}
\usage{
formatJobArgumentsHTML(job_args)
}
\arguments{
\item{job_args}{a list of job arguments, e.g. as read from the job_args.yml file}
}
\value{
a list of HTML-formatted key/value pairs
}
\description{
Format job arguments into html-formatted key/value pairs, for including
in an email
}
\examples{
\dontrun{
formatJobArgumentsHTML("/data/scratch/janani/molevolvr_out/Ba5sV1_full")
}
}
