% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineage.R
\name{GCA2Lineage}
\alias{GCA2Lineage}
\title{Function to map GCA_ID to TaxID, and TaxID to Lineage}
\usage{
GCA2Lineage(
  prot_data,
  assembly_path = "/data/research/jravilab/common_data/assembly_summary_genbank.txt",
  lineagelookup_path = "/data/research/jravilab/common_data/lineage_lookup.tsv",
  acc_col = "AccNum"
)
}
\arguments{
\item{prot_data}{Dataframe containing a column \code{GCA_ID}}

\item{assembly_path}{String of the path to the assembly_summary path
This file can be generated using the "downloadAssemblySummary()" function}

\item{lineagelookup_path}{String of the path to the lineage lookup file
(taxid to lineage mapping). This file can be generated using the
"create_lineage_lookup()" function}

\item{acc_col}{}
}
\description{
Function to map GCA_ID to TaxID, and TaxID to Lineage
}
\note{
Currently configured to have at most kingdom and phylum
}
\author{
Samuel Chen, Janani Ravi
}
