% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{generate_fa2tre}
\alias{generate_fa2tre}
\title{generate_fa2tre}
\usage{
generate_fa2tre(
  fa_file = "data/alns/pspa_snf7.fa",
  out_file = "data/alns/pspa_snf7.tre"
)
}
\arguments{
\item{fa_file}{Character. Path to file.
Default is 'pspa_snf7.fa'}

\item{out_file}{}
}
\description{
Generating phylogenetic tree from alignment file '.fa'
}
\details{
The alignment file would need two columns: 1. accession +
number and 2. alignment. The protein homolog accession to lineage mapping +
file should have
}
\note{
Please refer to the source code if you have alternate +
file formats and/or column names.
}
\examples{
\dontrun{
generate_aln2tree('pspa_snf7.fa')
}
}
\author{
Janani Ravi, MolEcologist
}
\keyword{alignment,}
\keyword{fasta}
\keyword{phylogenetic}
\keyword{tree,}
