% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msa.R
\name{generate_msa}
\alias{generate_msa}
\title{Function to generate MSA using kalign}
\usage{
generate_msa(fa_file = "", outfile = "")
}
\arguments{
\item{outfile}{}
}
\description{
Function to generate MSA using kalign
}
