% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{generate_trees}
\alias{generate_trees}
\title{generate_trees}
\usage{
generate_trees(aln_path = here("data/alns/"))
}
\arguments{
\item{aln_path}{}
}
\description{
Generate Trees for ALL fasta files in "data/alns"
}
