% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHANGED-pre-msa-tree.R
\name{generateAllAlignments2FA}
\alias{generateAllAlignments2FA}
\title{Adding Leaves to an alignment file w/ accessions}
\usage{
generateAllAlignments2FA(
  aln_path = here("data/rawdata_aln/"),
  fa_outpath = here("data/alns/"),
  lin_file = here("data/rawdata_tsv/all_semiclean.txt"),
  reduced = F
)
}
\arguments{
\item{aln_path}{Character. Path to alignment files.
Default is 'here("data/rawdata_aln/")'}

\item{fa_outpath}{Character. Path to file. Master protein file with AccNum & lineages.
Default is 'here("data/rawdata_tsv/all_semiclean.txt")'}

\item{lin_file}{Character. Path to the written fasta file.
Default is 'here("data/alns/")'.}

\item{reduced}{Boolean. If TRUE, the fasta file will contain only one sequence per lineage.
Default is 'FALSE'.}
}
\description{
Adding Leaves to all alignment files w/ accessions & DAs?
}
\details{
The alignment files would need two columns separated by spaces: 1. AccNum and 2. alignment. The protein homolog file should have AccNum, Species, Lineages.
}
\note{
Please refer to the source code if you have alternate + file formats and/or column names.
}
\examples{
\dontrun{
generateAllAlignments2FA()
}
}
\keyword{accnum,}
\keyword{alignment,}
\keyword{leaves,}
\keyword{lineage,}
\keyword{species}
