% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa2domain.R
\name{get_df_ipr_col_names}
\alias{get_df_ipr_col_names}
\title{Constructor function for interproscan column names
(based upon the global variable written in
molevol_scripts/R/colnames_molevol.R)}
\usage{
get_df_ipr_col_names()
}
\value{
\link{chr} interproscan column names used throughout molevolvr
}
\description{
Constructor function for interproscan column names
(based upon the global variable written in
molevol_scripts/R/colnames_molevol.R)
}
