% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa2domain.R
\name{get_df_ipr_col_types}
\alias{get_df_ipr_col_types}
\title{construct column types for reading interproscan output TSVs
(based upon the global variable written in
molevol_scripts/R/colnames_molevol.R)}
\usage{
get_df_ipr_col_types()
}
\value{
\link{collector} a named vector of type expecatations
for interproscan columns
}
\description{
construct column types for reading interproscan output TSVs
(based upon the global variable written in
molevol_scripts/R/colnames_molevol.R)
}
