% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_status_emails.R
\name{get_job_message}
\alias{get_job_message}
\title{Produces a mail message that can be sent to a user when their job is accepted.
Used by the send_job_status_email() method.}
\usage{
get_job_message(job_dir, pin_id, job_results_url, event_type, context)
}
\arguments{
\item{job_dir}{the directory where the job's arguments are stored, in job_args.yml}

\item{pin_id}{the unique identifier for the job}

\item{job_results_url}{the URL where the user can check the status of their job}

\item{event_type}{either 'start' or 'end', returns the corresponding email for the given type}

\item{context}{a list of additional values, e.g. job runtime info, that can be used in the template emails}
}
\value{
the result of the sendmailR::sendmail() call
}
\description{
Produces a mail message that can be sent to a user when their job is accepted.
Used by the send_job_status_email() method.
}
