% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_job_queue.R
\name{get_proc_weights}
\alias{get_proc_weights}
\title{Quickly get the runtime weights for MolEvolvR backend processes}
\usage{
get_proc_weights(medians_yml_path = NULL)
}
\arguments{
\item{dir_job_results}{\link{chr} path to MolEvolvR job_results
directory}
}
\value{
\link{list} names: processes; values: median runtime (seconds)

example: get_proc_weights()
}
\description{
Quickly get the runtime weights for MolEvolvR backend processes
}
