% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{get_process_q3s}
\alias{get_process_q3s}
\title{given a MolEvolvR log dataframe, calculate q3s for process runtimes}
\usage{
get_process_q3s(
  df_log,
  processes = c("dblast", "acc2info", "dblast_cleanup", "acc2fa", "blast_clust",
    "clust2table", "iprscan", "ipr2lineage", "ipr2da", "cln_blast")
)
}
\arguments{
\item{df_log}{\code{df_log} element from the return list of \code{aggregate_logs()}}
}
\value{
if group_by is empty: a single value; else: tibble from summarise
}
\description{
given a MolEvolvR log dataframe, calculate q3s for process runtimes
}
