% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa2domain.R
\name{getDomainsFromFA}
\alias{getDomainsFromFA}
\title{getDomainsFromFA}
\usage{
getDomainsFromFA(
  fasta,
  df_iprscan,
  analysis = c("Pfam", "Gene3D"),
  verbose = FALSE
)
}
\arguments{
\item{fasta}{\link{AAStringSet} a protein (AA) fasta}

\item{df_iprscan}{\link{tbl_df} the interproscan results from the original fasta}

\item{analysis}{\link{chr} the domain databases to extract sequences from}
}
\value{
fasta_domains \link{AAStringSet} fasta of domains
}
\description{
getDomainsFromFA
}
\examples{
\dontrun{
path_molevol_scripts <- file.path(Sys.getenv("DEV", unset = "/data/molevolvr_transfer/molevolvr_dev"), "molevol_scripts")
setwd(path_molevol_scripts)
source("R/fa2domain.R")
fasta <- Biostrings::readAAStringSet("./tests/example_protein.fa")
df_iprscan <- readIPRScanTSV("./tests/example_iprscan_valid.tsv")
getDomainsFromFA(fasta, df_iprscan)
}

}
