% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa2domain.R
\name{getIPRScanColNames}
\alias{getIPRScanColNames}
\title{Constructor function for interproscan column names
(based upon the global variable written in
molevol_scripts/R/colnames_molevol.R)}
\usage{
getIPRScanColNames()
}
\value{
\link[rlang:vector-construction]{rlang::chr} interproscan column names used throughout molevolvr
}
\description{
Constructor function for interproscan column names
(based upon the global variable written in
molevol_scripts/R/colnames_molevol.R)
}
