% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa2domain.R
\name{getIPRScanColTypes}
\alias{getIPRScanColTypes}
\title{construct column types for reading interproscan output TSVs
(based upon the global variable written in
molevol_scripts/R/colnames_molevol.R)}
\usage{
getIPRScanColTypes()
}
\value{
\link{collector} a named vector of type expecatations
for interproscan columns
}
\description{
construct column types for reading interproscan output TSVs
(based upon the global variable written in
molevol_scripts/R/colnames_molevol.R)
}
