% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accounting_db.R
\name{getJobsForCode}
\alias{getJobsForCode}
\title{getJobsForCode}
\usage{
getJobsForCode(code, cols = NA)
}
\arguments{
\item{code}{A job code to search for in job names}
}
\value{
Data frame with 'cols' columns (or all columns if 'cols') of jobs
}
\description{
Given a job code, retrieves SLURM jobs where that code occurs in their name
(i.e., the "job_name" column). Optionally, you may specify a vector of column
names to return; if unspecified, returns all columns.
}
