% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accounting_db.R
\name{getJobsWithStates}
\alias{getJobsWithStates}
\title{Get jobs with a specific SLURM job state.}
\usage{
getJobsWithStates(states = NA, cols = NA)
}
\arguments{
\item{states}{integers for states to search for; you can specify either a single
integer or a vector. if unspecified, returns all rows.}

\item{cols}{Column names to return; if unspecified, returns all columns.}
}
\value{
Data frame with 'cols' columns (or all columns if 'cols') of jobs
with the specified states, or all jobs if 'states' is unspecified
}
\description{
States are stored as integers, whose values derive from the file
https://github.com/SchedMD/slurm/tree/master/slurm/slurm.h. The possible
values and enumeration values are as follows.
\itemize{
\item 0:  JOB_PENDING: queued waiting for initiation
\item 1:  JOB_RUNNING: allocated resources and executing
\item 2:  JOB_SUSPENDED: allocated resources, execution suspended
\item 3:  JOB_COMPLETE: completed execution successfully
\item 4:  JOB_CANCELLED: cancelled by user
\item 5:  JOB_FAILED: completed execution unsuccessfully
\item 6:  JOB_TIMEOUT: terminated on reaching time limit
\item 7:  JOB_NODE_FAIL: terminated on node failure
\item 8:  JOB_PREEMPTED: terminated due to preemption
\item 9:  JOB_BOOT_FAIL: terminated due to node boot failure
\item 10: JOB_DEADLINE: terminated on deadline
\item 11: JOB_OOM: experienced out of memory error
}
(See https://slurm.schedmd.com/squeue.html#SECTION_JOB-STATE-CODES for
the meanings of these codes.)
}
\details{
For example, to search for all running, suspended, and pending jobs, you'd
provide the states argument states=c(0, 1, 2). If you wanted just running
jobs, you'd specify states=1.
}
