% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipr2viz.R
\name{getTopAccByLinDomArch}
\alias{getTopAccByLinDomArch}
\title{Group by lineage + DA then take top 20}
\usage{
getTopAccByLinDomArch(
  infile_full,
  DA_col = "DomArch.Pfam",
  lin_col = "Lineage_short",
  n = 20,
  query
)
}
\arguments{
\item{query}{}
}
\description{
Group by lineage + DA then take top 20
}
