% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acc2lin.R
\name{ipg2lin}
\alias{ipg2lin}
\title{ipg2lin}
\usage{
ipg2lin(accessions, ipg_file, assembly_path, lineagelookup_path)
}
\arguments{
\item{accessions}{Character vector of protein accessions}

\item{ipg_file}{Filepath to the file containing results of an efetch run on the
ipg database. The protein accession in 'accessions' should be contained in this
file}

\item{assembly_path}{String of the path to the assembly_summary path
This file can be generated using the "DownloadAssemblySummary()" function}

\item{lineagelookup_path}{String of the path to the lineage lookup file
(taxid to lineage mapping). This file can be generated using the
"create_lineage_lookup()" function}
}
\value{
Describe return, in detail
}
\description{
Takes the resulting file of an efetch run on the ipg database and
}
\examples{
\dontrun{
ipg2lin()
}

}
\author{
Samuel Chen, Janani Ravi
}
