% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipr2viz.R
\name{ipr2viz_web}
\alias{ipr2viz_web}
\title{IPR2Viz Web}
\usage{
ipr2viz_web(
  infile_ipr,
  accessions,
  analysis = c("Pfam", "Phobius", "TMHMM", "Gene3D"),
  group_by = "Analysis",
  name = "Name",
  text_size = 15,
  legend_name = "ShortName",
  cols = 5,
  rows = 10
)
}
\arguments{
\item{rows}{}
}
\description{
IPR2Viz Web
}
