% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{lineage.domain_repeats.plot}
\alias{lineage.domain_repeats.plot}
\title{Lineage Domain Repeats Plot}
\usage{
lineage.domain_repeats.plot(query_data, colname)
}
\arguments{
\item{colname}{}
}
\description{
Lineage Domain Repeats Plot
}
\examples{
\dontrun{
lineage.domain_repeats.plot()
}
}
