% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{lineage.neighbors.plot}
\alias{lineage.neighbors.plot}
\title{Lineage Plot for top neighbors}
\usage{
lineage.neighbors.plot(query_data = "prot", query = "pspa",
  colname = "GenContext.norep")
}
\arguments{
\item{query_data}{Data frame of protein homologs with the usual 11 columns +
additional word columns (0/1 format). Default is pspa_data.}

\item{query}{Name of query protein/domain. Default is "pspa".}

\item{colname}{Column name from query_data. Default is "GenContext.norep".}
}
\description{
Lineage plot for top neighbors obtained from DAs of
Genomic Contexts.
}
\details{
For "da2doms" you would need the file DA.doms.wc as well as the
column query_data$DomArch.norep

For "gc2da", you would need the file GC.DA.wc as well as the column
query_data$GenContext.norep
}
\note{
Please refer to the source code if you have alternate file formats and/or
column names.
}
\examples{
\dontrun{
lineage.neighbors.plot(pspa_data, pspa, "GenContext.norep", "da2doms")
}

}
\author{
Janani Ravi
}
\keyword{Architectures,}
\keyword{Domain}
\keyword{Domains,}
\keyword{GenomicContexts}
\keyword{Lineages,}
