% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanup.R
\name{make_accnums_unique}
\alias{make_accnums_unique}
\title{make accnums unique}
\usage{
make_accnums_unique(accnums)
}
\arguments{
\item{accnums}{\link{chr} a vector of accession numbers}
}
\value{
\link{chr} a vector of adjusted, unique accession numbers
}
\description{
Append an index of occurence suffix to each accession number (or any
character vector) making them unique
}
\examples{
\dontrun{
c("xxx", "xxx", "xxx", "yyy", "yyy") |>
    make_accnums_unique()
}
}
