% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_status_emails.R
\name{make_job_results_url}
\alias{make_job_results_url}
\title{Given a pin_id, returns the URL where the user can check the status of their job}
\usage{
make_job_results_url(
  pin_id,
  base_url = Sys.getenv("BASE_URL", unset = "http://jravilab.org/molevolvr/")
)
}
\arguments{
\item{pin_id}{the unique identifier for the job}

\item{base_url}{the base URL for the MolEvolvR web app}
}
\value{
the URL where the user can check the status of their job
}
\description{
The 'base_url' parameter is set from the environment variable 'BASE_URL' if it's
available, and defaults to "http://jravilab.org/molevolvr/" if it's not.
}
