% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_job_queue.R
\name{make_opts2procs}
\alias{make_opts2procs}
\title{Construct list where names (MolEvolvR advanced options) point to processes}
\usage{
make_opts2procs()
}
\value{
list where names (MolEvolvR advanced options) point to processes

example: list_opts2procs <- make_opts2procs
}
\description{
Construct list where names (MolEvolvR advanced options) point to processes
}
