% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_job_queue.R
\name{map_advanced_opts2procs}
\alias{map_advanced_opts2procs}
\title{Use MolEvolvR advanced options to get associated processes}
\usage{
map_advanced_opts2procs(advanced_opts)
}
\arguments{
\item{advanced_opts}{character vector of MolEvolvR advanced options}
}
\value{
character vector of process names that will execute given
the advanced options

example:
advanced_opts <- c("homology_search", "domain_architecture")
procs <- map_advanced_opts2procs(advanced_opts)
}
\description{
Use MolEvolvR advanced options to get associated processes
}
