% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHANGED-pre-msa-tree.R, R/pre-msa-tree.R
\name{mapAcc2Name}
\alias{mapAcc2Name}
\title{mapAcc2Name}
\usage{
mapAcc2Name(line, acc2name, acc_col = "AccNum", name_col = "Name")

mapAcc2Name(line, acc2name, acc_col = "AccNum", name_col = "Name")
}
\arguments{
\item{line}{The line of a fasta file starting with '>'}

\item{acc2name}{Data Table containing a column of accession numbers and a name column}

\item{acc_col}{Name of the column containing Accession numbers}

\item{name_col}{Name of the column containing the names that the accession numbers
are mapped to}
}
\value{
A character string representing the updated FASTA line, where the
accession number is replaced with its corresponding name.

Character string. The modified line from the Fasta file header with
the name instead of the accession number.
}
\description{
Default renameFA() replacement function. Maps an accession number to its name

Default rename_fasta() replacement function. Maps an accession number to its name
}
\examples{
\dontrun{
mapAcc2Name(">P12345 some description", acc2name, "AccNum", "Name")
}
\dontrun{
acc2name_table <- data.table(AccNum = c("ACC001", "ACC002"),
Name = c("Species A", "Species B"))
line <- ">ACC001 some additional info"
mapped_line <- mapAcc2Name(line, acc2name_table)
mapped_line  # Expected output: ">Species A"
}
}
