% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_job_queue.R
\name{mapAdvOption2Process}
\alias{mapAdvOption2Process}
\title{mapAdvOption2Process}
\usage{
mapAdvOption2Process(advanced_opts)
}
\arguments{
\item{advanced_opts}{character vector of MolEvolvR advanced options}
}
\value{
character vector of process names that will execute given
the advanced options

example:
advanced_opts <- c("homology_search", "domain_architecture")
procs <- mapAdvOption2Process(advanced_opts)
}
\description{
Use MolEvolvR advanced options to get associated processes
}
