% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_job_queue.R
\name{mapOption2Process}
\alias{mapOption2Process}
\title{mapOption2Process}
\usage{
mapOption2Process()
}
\value{
list where names (MolEvolvR advanced options) point to processes

example: list_opts2procs <- mapOption2Process
}
\description{
Construct list where names (MolEvolvR advanced options) point to processes
}
