% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{MolEvolvR_summary}
\alias{MolEvolvR_summary}
\alias{summarizeByLineage}
\alias{summarizeDomArch_ByLineage}
\alias{summarizeDomArch}
\alias{summarizeGenContext_ByDomArchLineage}
\alias{summarizeGenContext_ByLineage}
\alias{summarizeGenContext}
\alias{totalGenContextOrDomArchCounts}
\title{MolEvolvR Summary}
\usage{
summarizeByLineage(prot = "prot", column = "DomArch", by = "Lineage", query)

summarizeDomArch_ByLineage(x)

summarizeDomArch(x)

summarizeGenContext_ByDomArchLineage(x)

summarizeGenContext_ByLineage(x)

summarizeGenContext(x)

totalGenContextOrDomArchCounts(
  prot,
  column = "DomArch",
  lineage_col = "Lineage",
  cutoff = 90,
  RowsCutoff = FALSE,
  digits = 2
)
}
\arguments{
\item{prot}{A data frame that must contain columns:
\itemize{\item Either 'GenContext' or 'DomArch.norep' \item count}}

\item{column}{Character. The column to summarize, default is "DomArch".}

\item{by}{A string representing the grouping column (e.g., \code{Lineage}).
Default is "Lineage".}

\item{query}{A string specifying the query pattern for filtering the target
column. Use "all" to skip filtering and include all rows.}

\item{x}{A dataframe or tibble containing the data. It must have columns
named \code{GenContext}, \code{DomArch}, and \code{Lineage}.}

\item{lineage_col}{Character. The name of the lineage column, default is
"Lineage".}

\item{cutoff}{Numeric. Cutoff for total count. Counts below this cutoff value
will not be shown. Default is 0.}

\item{RowsCutoff}{Logical. If TRUE, filters based on cumulative percentage
cutoff. Default is FALSE.}

\item{digits}{Numeric. Number of decimal places for percentage columns.
Default is 2.}
}
\value{
A tibble summarizing the counts of occurrences of elements in
the \code{column}, grouped by the \code{by} column. The result includes the number
of occurrences (\code{count}) and is arranged in descending order of count.

A tibble summarizing the counts of unique domain architectures
(\code{DomArch}) per lineage (\code{Lineage}). The resulting table contains three
columns: \code{DomArch}, \code{Lineage}, and \code{count}, which indicates the frequency
of each domain architecture for each lineage. The results are arranged in
descending order of \code{count}.

A tibble summarizing each unique \code{DomArch}, along with the following
columns:
\itemize{
\item \code{totalcount}: The total occurrences of each \code{DomArch} across all lineages.
\item \code{totallin}: The total number of unique lineages in which each \code{DomArch}
appears.
The results are arranged in descending order of \code{totallin} and \code{totalcount}.
}

A tibble summarizing each unique combination of \code{GenContext},
\code{DomArch}, and \code{Lineage}, along with the following columns:
\itemize{
\item \code{GenContext}: The genomic context for each entry.
\item \code{DomArch}: The domain architecture for each entry.
\item \code{Lineage}: The lineage associated with each entry.
\item \code{count}: The total number of occurrences for each combination of
\code{GenContext}, \code{DomArch}, and \code{Lineage}.
}

The results are arranged in descending order of \code{count}.

Describe return, in detail

A tibble summarizing each unique combination of \code{GenContext} and
\code{Lineage}, along with the following columns:
\itemize{
\item \code{GenContext}: The genomic context for each entry.
\item \code{Lineage}: The lineage associated with each entry.
\item \code{count}: The total number of occurrences for each combination of
\code{GenContext} and \code{Lineage}.
}

The results are arranged in descending order of \code{count}.

A data frame with the following columns:
\itemize{
\item \code{{{ column }}}: Unique values from the specified column.
\item \code{totalcount}: The total count of occurrences for each unique value in
the specified column.
\item \code{IndividualCountPercent}: The percentage of each \code{totalcount} relative to
the overall count.
\item \code{CumulativePercent}: The cumulative percentage of total counts.
}
}
\description{
A collection of summary functions for the MolEvolvR package.

Function to summarize and retrieve counts by Domains & Domains+Lineage

Function to retrieve counts of how many lineages a DomArch appears in

Creates a data frame with a totalcount column

This function is designed to sum the counts column by either Genomic Context or Domain Architecture and creates a totalcount column from those sums.
}
\note{
Please refer to the source code if you have alternate file formats and/or
column names.
}
\examples{
\dontrun{
library(tidyverse)
tibble(DomArch = c("a+b", "a+b", "b+c", "a+b"), Lineage = c("l1", "l1", "l1", "l2")) |>
    summarizeByLineage(query = "all")
}

\dontrun{
summarizeDomArch_ByLineage()
}
\dontrun{
summarizeDomArch()
}
\dontrun{
summarizeGenContext_ByDomArchLineage
}
\dontrun{
summarizeGenContext_ByLineage()
}
\dontrun{
summarizeGenContext()
}
\dontrun{
totalGenContextOrDomArchCounts(pspa - gc_lin_counts, 0, "GC")
}
}
