% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msa.R
\name{msa_pdf}
\alias{msa_pdf}
\title{Multiple Sequence Alignment}
\usage{
msa_pdf(fasta_path, out_path = NULL, lowerbound = NULL, upperbound = NULL)
}
\arguments{
\item{fasta_path}{Character. The path location of the fasta file to be read.}

\item{out_path}{Character. The path location of the output pdf to write.
Default is NULL. If value is NULL, the pdf is written to the same directory as the fasta file.}

\item{lowerbound}{Numeric. The column that determines the starting location of the MSA.
Default is NULL. If value is NULL, the entire multiple sequence alignment is printed.}

\item{upperbound}{Numeric. The column that determines the ending location of the MSA.
Default is NULL. If value is NULL, the entire multiple sequence alignment is printed.}
}
\description{
Generates a multiple sequence alignment from a fasta file

msa_pdf is a function that reads a fasta file and generates a multiple sequence alignment as
a pdf
}
\examples{
\dontrun{
msa_pdf()
}
}
