% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanup.R
\name{pick_longer_duplicate}
\alias{pick_longer_duplicate}
\title{Pick Longer Duplicate}
\usage{
pick_longer_duplicate(prot, column)
}
\arguments{
\item{column}{}
}
\value{
Describe return, in detail
}
\description{
Pick Longer Duplicate
}
\examples{
\dontrun{
pick_longer_duplicate()
}
}
