% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_job_queue.R
\name{plot_estimated_walltimes}
\alias{plot_estimated_walltimes}
\title{Plot the estimated runtimes for different advanced options and number
of inputs}
\usage{
plot_estimated_walltimes()
}
\value{
line plot object

example:
p <- plot_estimated_walltimes()
ggplot2::ggsave(filename = "/data/molevolvr_transfer/molevolvr_
dev/molevol_scripts/docs/estimate_walltimes.png", plot = p)
}
\description{
this function was just for fun; very, very messy code
}
