% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InLP_from_molevol_scripts.R, R/plotting.R
\name{plotDomarchTreeMSA}
\alias{plotDomarchTreeMSA}
\title{Plots the interproscan visualization aligned to a tree with inline MSA plots}
\usage{
plotDomarchTreeMSA(
  app_data,
  rep,
  phyloSelect,
  tree_msa_tool,
  rval_phylo,
  acc_to_name,
  analysis = c("Pfam"),
  show.tiplabels = FALSE,
  progress = function(...) {
 },
  verbose = TRUE
)

plotDomarchTreeMSA(
  app_data,
  rep,
  phyloSelect,
  tree_msa_tool,
  rval_phylo,
  acc_to_name,
  analysis = c("Pfam"),
  show.tiplabels = FALSE,
  progress = function(...) {
 },
  verbose = TRUE
)
}
\arguments{
\item{app_data}{The app data object containing the input data.}

\item{rep}{The representative sequences.}

\item{phyloSelect}{The selected protein on the phylogeny page.}

\item{tree_msa_tool}{The selected MSA tool on the phylogeny page (e.g., Clustal Omega, Clustal W, Muscle).}

\item{rval_phylo}{Indicates whether the advanced option "Phylogenetic Analysis (for homologs)" was selected for the job.}

\item{acc_to_name}{A mapping of accessions to names.}

\item{analysis}{The analyses to display in the iprscan plot.}

\item{show.tiplabels}{Whether to show tip labels on the tree plot.}

\item{progress}{Optional, function invoked to report progress to the caller.}

\item{verbose}{Optional, whether to print verbose messages.}
}
\value{
A combined plot of the phylogenetic tree and interproscan visualization.

A combined plot of the phylogenetic tree and interproscan visualization.
}
\description{
Plots the interproscan visualization aligned to a tree with inline MSA plots

Plots the interproscan visualization aligned to a tree with inline MSA plots
}
