% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_job_queue.R
\name{plotEstimatedWallTimes}
\alias{plotEstimatedWallTimes}
\title{plotEstimatedWallTimes}
\usage{
plotEstimatedWallTimes()
}
\value{
line plot object

example:
p <- plotEstimatedWallTimes()
ggplot2::ggsave(filename = "/data/molevolvr_transfer/molevolvr_
dev/molevol_scripts/docs/estimate_walltimes.png", plot = p)
}
\description{
Plot the estimated runtimes for different advanced options and number
of inputs

this function was just for fun; very, very messy code
}
