% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipr2viz.R
\name{plotIPR2Viz}
\alias{plotIPR2Viz}
\title{IPR2Viz}
\usage{
plotIPR2Viz(
  infile_ipr = NULL,
  infile_full = NULL,
  accessions = c(),
  analysis = c("Pfam", "Phobius", "TMHMM", "Gene3D"),
  group_by = "Analysis",
  topn = 20,
  name = "Name",
  text_size = 15,
  query = "All"
)
}
\arguments{
\item{query}{}
}
\description{
IPR2Viz
}
