% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotLineageDA}
\alias{plotLineageDA}
\title{Lineage Plot: Heatmap of Domains/DAs/GCs vs Lineages}
\usage{
plotLineageDA(
  query_data = "prot",
  colname = "DomArch",
  cutoff = 90,
  RowsCutoff = FALSE,
  color = "default"
)
}
\arguments{
\item{query_data}{Data frame of protein homologs with the usual 11 columns +
additional word columns (0/1 format).
Default is prot (variable w/ protein data).}

\item{colname}{Column name from query_data: "DomArch.norep", "GenContext.norep",
"DomArch.PFAM.norep" or "DomArch.LADB.norep". Default is "DomArch.norep".}

\item{cutoff}{Numeric. Cutoff for word frequency. Default is 90.}

\item{RowsCutoff}{Boolean. If TRUE, applies a row cutoff to remove data rows
based on a certain condition. Default is FALSE.}

\item{color}{Color for the heatmap. One of six options: "default", "magma", "inferno",
"plasma", "viridis", or "cividis"}
}
\value{
A LineageDA plot object.
}
\description{
Lineage plot for Domains, Domain Architectures and
Genomic Contexts. Heatmap.
}
\details{
For "da2doms" you would need the file DA.doms.wc as well as the
column query_data$DomArch.norep

For "gc2da", you would need the file GC.DA.wc as well as the column
query_data$GenContext.norep
}
\note{
Please refer to the source code if you have alternate file formats and/or
column names.
}
\examples{
\dontrun{
plotLineageDA(toast_rack_data, 10, "DomArch.norep", "da2doms")
}
}
\author{
Janani Ravi
}
\keyword{Architectures,}
\keyword{Domain}
\keyword{Domains,}
\keyword{GenomicContexts}
\keyword{Lineages,}
