% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotLineageDomainRepeats}
\alias{plotLineageDomainRepeats}
\title{Lineage Domain Repeats Plot}
\usage{
plotLineageDomainRepeats(query_data, colname)
}
\arguments{
\item{colname}{}
}
\description{
Lineage Domain Repeats Plot
}
\examples{
\dontrun{
plotLineageDomainRepeats()
}
}
