% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotLineageHeatmap}
\alias{plotLineageHeatmap}
\title{plotLineageHeatmap}
\usage{
plotLineageHeatmap(prot, domains_of_interest, level = 3, label.size = 8)
}
\arguments{
\item{prot}{Data frame containing DomArch and Lineage Columns}

\item{domains_of_interest}{Vector of domains to check for the presence of in all the lineages}

\item{level}{The max depth of Lineage. ie) i = Kingdom, 2 = Phylum, 3 = class ...}

\item{label.size}{Size of the text labels}
}
\description{
Generate a lineage plot
}
\examples{
\dontrun{
plotLineageHeatmap(psp_data,
    c(
        "PspA", "Snf7", "Classical-AAA", "PspF", "PspB", "PspC", "ClgR", "PspM",
        "Thioredoxin", "PspN_N", "DUF3046", "LiaI-LiaF-TM", "Toast_rack", "REC",
        "HISKIN", "HAAS", "SHOCT-bihelical", "SHOCT-like", "Tfu_1009", "PspAA",
        "Spermine_synth", "TM-Flotillin", "Band-7", "Betapropeller",
        "MacB_PCD", "FTSW_RODA_SPOVE", "Cest_Tir", "SIGMA-HTH", "GNTR-HTH",
        "DUF2089-HTH", "PadR-HTH", "RHH", "ZnR"
    ),
    level = 2
)
}

}
\author{
Samuel Chen, Janani Ravi
}
