% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotLineageQuery}
\alias{plotLineageQuery}
\title{Lineage Plot: Heatmap of Queries vs Lineages}
\usage{
plotLineageQuery(
  query_data = all,
  queries,
  colname = "ClustName",
  cutoff,
  color = "default"
)
}
\arguments{
\item{query_data}{Data frame of protein homologs with the usual 11 columns +
additional word columns (0/1 format).
Default is prot (variable w/ protein data).}

\item{queries}{Character Vector containing the queries that will be used for
the categories.}

\item{colname}{Character. The column used for filtering based on the \code{queries}.
Default is "ClustName".}

\item{cutoff}{Numeric. The cutoff value for filtering rows based on their
total count. Rows with values below this cutoff are excluded.}

\item{color}{Character. Defines the color palette used for the heatmap.
Default is a red gradient.}
}
\value{
A ggplot object representing a heatmap (tile plot) showing the
relationship between queries and lineages, with the intensity of color
representing the count of matching records.
}
\description{
Lineage plot for queries. Heatmap.
}
\note{
Please refer to the source code if you have alternate file formats and/or
column names.
}
\examples{
\dontrun{
plotLineageQuery(prot, c("PspA", "PspB", "PspC", "PspM", "PspN"), 95)
}
}
\author{
Janani Ravi, Samuel Chen
}
\keyword{Architectures,}
\keyword{Domain}
\keyword{Domains,}
\keyword{GenomicContexts}
\keyword{Lineages,}
