% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotme.R
\name{plotSunburst}
\alias{plotSunburst}
\alias{plotTreemap}
\title{Create an interactive plotly from count data}
\usage{
plotSunburst(count_data, fill_by_n = FALSE, sort_by_n = FALSE, maxdepth = 2)

plotTreemap(count_data, fill_by_n = FALSE, sort_by_n = FALSE)
}
\arguments{
\item{count_data}{}

\item{fill_by_n}{If TRUE, uses a continuous scale to fill plot by group size}

\item{sort_by_n}{}
}
\description{
These functions help you quickly create interactive hierarchical plots
from categorical data. They expect the summary of the data created by
\code{dplyr::count()} and produce either a sunburst plot (\code{plotSunburst()}) or
a treemap plot (\code{plotTreemap()})
}
\examples{
library(dplyr)
starwars_count <- count(starwars, species, eye_color, name)

# sunburst plot
plotSunburst(starwars_count)

# fill by group size
plotSunburst(starwars_count, fill_by_n = TRUE)

# treemap plot, ordered by group size
plotTreemap(starwars_count, sort_by_n = TRUE)

# display al charchaters by homeworld
starwars \%>\%
    count(homeworld, name) \%>\%
    plotTreemap(sort_by_n = TRUE)

}
