% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotme.R
\name{prepareColumnParams}
\alias{prepareColumnParams}
\title{prepareColumnParams}
\usage{
prepareColumnParams(count_data, fill_by_n, sort_by_n)
}
\arguments{
\item{sort_by_n}{}
}
\description{
prepareColumnParams
}
