% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotme.R
\name{prepareSingleColumnParams}
\alias{prepareSingleColumnParams}
\title{prepareSingleColumnParams}
\usage{
prepareSingleColumnParams(df, col_num, root)
}
\arguments{
\item{root}{}
}
\description{
prepareSingleColumnParams
}
