% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineage.R
\name{proteinAcc2TaxID}
\alias{proteinAcc2TaxID}
\title{proteinAcc2TaxID}
\usage{
proteinAcc2TaxID(accnums, suffix, out_path, return_dt = FALSE)
}
\arguments{
\item{return_dt}{}
}
\description{
proteinAcc2TaxID
}
