% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineage.R
\name{proteinAcc2TaxID_old}
\alias{proteinAcc2TaxID_old}
\title{proteinAcc2TaxID_old}
\usage{
proteinAcc2TaxID_old(accessions, out_path, plan = "multicore")
}
\arguments{
\item{accessions}{A character vector containing the accession numbers to query
in the protein database.}

\item{out_path}{A string specifying the path where the results of the query
will be written. If set to NULL, a temporary directory will be used.}

\item{plan}{A character string that specifies the execution plan for parallel
processing. The default is "multicore".}
}
\value{
This function does not return a value. It writes the results to the
specified output path.
}
\description{
Perform elink to go from protein database to taxonomy database
and write the resulting file of taxid and lineage to out_path
}
\examples{
\dontrun{
accessions <- c("ABC123", "XYZ456", "LMN789")
proteinAcc2TaxID_old(accessions, out_path = "/path/to/output")
}
}
\author{
Samuel Chen, Janani Ravi
}
