% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa2domain.R
\name{read_iprscan_tsv}
\alias{read_iprscan_tsv}
\title{Read an interproscan output TSV with standardized
column names and types}
\usage{
read_iprscan_tsv(filepath)
}
\arguments{
\item{filepath}{\link{chr} path to interproscan output TSV}
}
\value{
\link{tbl_df} interproscan output table
}
\description{
Read an interproscan output TSV with standardized
column names and types
}
