% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanup.R
\name{remove_tails}
\alias{remove_tails}
\title{Remove Tails}
\usage{
remove_tails(prot, by_column = "DomArch", keep_domains = FALSE)
}
\arguments{
\item{prot}{A data frame containing 'DomArch', 'GenContext', 'ClustName' columns.}

\item{by_column}{Default column is 'DomArch'. Can also take 'ClustName', 'GenContext' as input.}

\item{keep_domains}{Default is False Keeps tail entries that contain the query domains.}
}
\value{
Describe return, in detail
}
\description{
Remove tails/singletons

This function ...
Certain low frequency domain architectures can be removed.
The original data frame is returned with the corresponding cleaned up column.
}
\examples{
\dontrun{
remove_tails(prot, "DomArch")
}
}
