% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanup.R
\name{removeAsterisks}
\alias{removeAsterisks}
\title{Remove Asterisk}
\usage{
removeAsterisks(query_data, colname = "GenContext")
}
\arguments{
\item{query_data}{A data frame containing the data to be processed.}

\item{colname}{The name of the column from which asterisks should be removed.
Defaults to "GenContext".}
}
\value{
The original data frame with asterisks removed from the specified
column.
}
\description{
Remove the asterisks from a column of data
Used for removing * from GenContext columns
}
\examples{
\dontrun{
removeAsterisks()
}
}
