% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanup.R
\name{removeEmptyRows}
\alias{removeEmptyRows}
\title{Remove Empty}
\usage{
removeEmptyRows(prot, by_column = "DomArch")
}
\arguments{
\item{prot}{A data frame containing 'DomArch', 'Species', 'GenContext', 'ClustName' columns.}

\item{by_column}{Column by which empty rows should be removed to domain+domain -> domain(s).
Default column is 'DomArch'. Can also take the following as input, 'Species', 'GenContext', 'ClustName'.}
}
\value{
A tibble with rows removed where the specified column contains
\code{"-"}, \code{"NA"}, or an empty string.
}
\description{
Remove empty rows by column

Removes empty rows in the specified column.

This function ...
The original data frame is returned with the corresponding cleaned up column.
}
\examples{
\dontrun{
removeEmptyRows(prot, "DomArch")
}
}
