% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanup.R
\name{replaceQMs}
\alias{replaceQMs}
\title{Replace QMs}
\usage{
replaceQMs(prot, by_column = "GenContext")
}
\arguments{
\item{prot}{DataTable to operate on}

\item{by_column}{Column to operate on}
}
\value{
Describe return, in detail
}
\description{
Replace consecutive '?' separated by '->', '<-' or '||' with 'X(s)'
Replace '?' with 'X'
}
\examples{
\dontrun{
replaceQMs()
}

}
