% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHANGED-pre-msa-tree.R, R/pre-msa-tree.R
\name{RepresentativeAccNums}
\alias{RepresentativeAccNums}
\title{Function to generate a vector of one Accession number per distinct observation from 'reduced' column}
\usage{
RepresentativeAccNums(prot_data, reduced = "Lineage", accnum_col = "AccNum")

RepresentativeAccNums(prot_data, reduced = "Lineage", accnum_col = "AccNum")
}
\arguments{
\item{prot_data}{Data frame containing Accession Numbers}

\item{reduced}{Column from prot_data from which distinct observations
will be generated from.
One accession number will be assigned for each of these observations}

\item{accnum_col}{Column from prot_data that contains Accession Numbers}
}
\description{
Function to generate a vector of one Accession number per distinct observation from 'reduced' column
}
\author{
Samuel Chen, Janani Ravi
}
