% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse_operons.R
\name{reverse_operons}
\alias{reverse_operons}
\title{Reverse Operons}
\usage{
reverse_operons(prot)
}
\arguments{
\item{prot}{A data frame containing a GenContext column}
}
\description{
Reverses the direction of the operons
}
\details{
This function reverses the direction of operons in a column of a data frame called 'GenContext'. Returns
the original dataframe with the original operons in a column called 'GenContext.old' and the reversed
operons in a column called 'GenContext'
}
\examples{
reverse_operons(pspa.sub)
}
