% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blastWrappers.R
\name{run_deltablast}
\alias{run_deltablast}
\title{Run DELTABLAST to find homologs for proteins of interest}
\usage{
run_deltablast(
  deltablast_path,
  db_search_path,
  db = "refseq",
  query,
  evalue = "1e-5",
  out,
  num_alignments,
  num_threads = 1
)
}
\arguments{
\item{db_search_path}{Path to the BLAST databases}
}
\description{
Run DELTABLAST to find homologs for proteins of interest
}
\author{
Samuel Chen, Janani Ravi
}
