% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanup.R
\name{selectLongestDuplicate}
\alias{selectLongestDuplicate}
\title{Pick Longer Duplicate}
\usage{
selectLongestDuplicate(prot, column)
}
\arguments{
\item{prot}{A data frame containing the data, with at least one column
named 'AccNum' for identification of duplicates.}

\item{column}{The name of the column from which the longest entry among
duplicates will be selected.}
}
\value{
A data frame containing only the longest entries among duplicates
based on the specified column.
}
\description{
Pick Longer Duplicate
}
\examples{
\dontrun{
selectLongestDuplicate()
}
}
