% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_status_emails.R
\name{send_job_status_email}
\alias{send_job_status_email}
\title{Sends a "job accepted" email to a user when their job is accepted,
including details about the job submission and how to check its status.}
\usage{
send_job_status_email(
  notify_email,
  job_dir,
  pin_id,
  event_type,
  context = NULL
)
}
\arguments{
\item{notify_email}{the email address to send the notification to}

\item{job_dir}{the directory where the job's arguments are stored, in job_args.yml}

\item{pin_id}{the unique identifier for the job}

\item{event_type}{either 'start' or 'end', returns the corresponding email for the given type}

\item{job_results_url}{the URL where the user can check the status of their job}
}
\value{
the result of the sendmailR::sendmail() call
}
\description{
Sends a "job accepted" email to a user when their job is accepted,
including details about the job submission and how to check its status.
}
