% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{shorten_lineage}
\alias{shorten_lineage}
\title{Shorten Lineage}
\usage{
shorten_lineage(data, colname = "Lineage", abr_len = 1)
}
\arguments{
\item{abr_len}{}
}
\description{
Shorten Lineage
}
\examples{
\dontrun{
shorten_lineage()
}
}
