% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{shortenLineage}
\alias{shortenLineage}
\title{Shorten Lineage}
\usage{
shortenLineage(data, colname = "Lineage", abr_len = 1)
}
\arguments{
\item{abr_len}{}
}
\description{
Shorten Lineage
}
\examples{
\dontrun{
shortenLineage()
}
}
