% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acc2lin.R
\name{sink.reset}
\alias{sink.reset}
\title{Sink Reset}
\usage{
sink.reset()
}
\value{
No return, but run to close all outstanding \code{sink()}s
}
\description{
Sink Reset
}
\examples{
\dontrun{
sink.reset()
}
}
