% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{stacked_lin_plot}
\alias{stacked_lin_plot}
\title{Stacked Lineage Plot}
\usage{
stacked_lin_plot(
  prot,
  column = "DomArch",
  cutoff,
  Lineage_col = "Lineage",
  xlabel = "Domain Architecture",
  reduce_lineage = TRUE,
  label.size = 8,
  legend.position = c(0.7, 0.4),
  legend.text.size = 10,
  legend.cols = 2,
  legend.size = 0.7,
  coord_flip = TRUE,
  legend = TRUE
)
}
\arguments{
\item{legend}{}
}
\description{
Stacked Lineage Plot
}
\examples{
\dontrun{
stacked_lin_plot()
}
}
