% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanup.R
\name{string2accnum}
\alias{string2accnum}
\title{string2accnum}
\usage{
string2accnum(string)
}
\arguments{
\item{string}{}
}
\value{
Describe return, in detail
}
\description{
string2accnum
}
\examples{
\dontrun{
string2accnum()
}
}
