% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summ.DA.byLin}
\alias{summ.DA.byLin}
\title{summ.DA.byLin}
\usage{
summ.DA.byLin(x)
}
\arguments{
\item{x}{A dataframe or tibble containing the data. It must have columns
named \code{DomArch} and \code{Lineage}.}
}
\value{
A tibble summarizing the counts of unique domain architectures
(\code{DomArch}) per lineage (\code{Lineage}). The resulting table contains three
columns: \code{DomArch}, \code{Lineage}, and \code{count}, which indicates the frequency
of each domain architecture for each lineage. The results are arranged in
descending order of \code{count}.
}
\description{
Function to summarize and retrieve counts by Domains & Domains+Lineage
}
\examples{
\dontrun{
summ.DA.byLin()
}
}
