% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summ.GC.byDALin}
\alias{summ.GC.byDALin}
\title{summ.GC.byDALin}
\usage{
summ.GC.byDALin(x)
}
\arguments{
\item{x}{}
}
\value{
Define return, in detail
}
\description{
summ.GC.byDALin
}
\examples{
\dontrun{
summ.GC.byDALin
}
}
