% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summ.GC.byLin}
\alias{summ.GC.byLin}
\title{summ.GC.byLin}
\usage{
summ.GC.byLin(x)
}
\arguments{
\item{x}{}
}
\value{
Describe return, in detail
}
\description{
summ.GC.byLin
}
\examples{
\dontrun{
summ.GC.byLin()
}
}
