% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summ.GC}
\alias{summ.GC}
\title{summ.GC}
\usage{
summ.GC(x)
}
\arguments{
\item{x}{A dataframe or tibble containing the data. It must have columns
named \code{GenContext}, \code{DomArch}, and \code{Lineage}.}
}
\value{
A tibble summarizing each unique combination of \code{GenContext} and
\code{Lineage}, along with the following columns:
\itemize{
\item \code{GenContext}: The genomic context for each entry.
\item \code{Lineage}: The lineage associated with each entry.
\item \code{count}: The total number of occurrences for each combination of
\code{GenContext} and \code{Lineage}.
}

The results are arranged in descending order of \code{count}.
}
\description{
summ.GC
}
\examples{
\dontrun{
summ.GC()
}
}
