% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summarize}
\alias{summarize}
\alias{filter_by_doms}
\alias{count_bycol}
\alias{elements2words}
\alias{words2wc}
\alias{filter_freq}
\alias{summarize_bylin}
\alias{summ.DA.byLin}
\alias{summ.DA}
\alias{summ.GC.byDALin}
\alias{summ.GC.byLin}
\alias{summ.GC}
\alias{total_counts}
\title{Filter by Domains}
\usage{
filter_by_doms(
  prot,
  column = "DomArch",
  doms_keep = c(),
  doms_remove = c(),
  ignore.case = FALSE
)

count_bycol(prot = prot, column = "DomArch", min.freq = 1)

elements2words(prot, column = "DomArch", conversion_type = "da2doms")

words2wc(string)

filter_freq(x, min.freq)

summarize_bylin(prot = "prot", column = "DomArch", by = "Lineage", query)

summ.DA.byLin(x)

summ.DA(x)

summ.GC.byDALin(x)

summ.GC.byLin(x)

summ.GC(x)

total_counts(
  prot,
  column = "DomArch",
  lineage_col = "Lineage",
  cutoff = 90,
  RowsCutoff = FALSE,
  digits = 2
)
}
\arguments{
\item{prot}{A data frame that must contain columns:
\itemize{\item Either 'GenContext' or 'DomArch.norep' \item count}}

\item{column}{Character. The column to summarize, default is "DomArch".}

\item{doms_keep}{Vector of domains that must be identified within column in order for
observation to be kept}

\item{doms_remove}{Vector of domains that, if found within an observation, will be removed}

\item{ignore.case}{Should the matching be non case sensitive}

\item{min.freq}{A numeric value specifying the minimum frequency threshold.
Only elements with frequencies greater than or equal to this value will be
retained.}

\item{conversion_type}{A character string specifying the type of conversion.
Two options are available:
\describe{
\item{\code{da2doms}}{Convert domain architectures into individual domains by
replacing \code{+} symbols with spaces.}
\item{\code{gc2da}}{Convert genomic context into domain architectures by
replacing directional symbols (\verb{<-}, \verb{->}, and \code{|}) with spaces.}
}}

\item{string}{A character string containing the elements (words) to count.
This would typically be a space-delimited string representing domain
architectures or genomic contexts.}

\item{x}{A dataframe or tibble containing the data. It must have columns
named \code{GenContext}, \code{DomArch}, and \code{Lineage}.}

\item{by}{A string representing the grouping column (e.g., \code{Lineage}).
Default is "Lineage".}

\item{query}{A string specifying the query pattern for filtering the target
column. Use "all" to skip filtering and include all rows.}

\item{lineage_col}{Character. The name of the lineage column, default is
"Lineage".}

\item{cutoff}{Numeric. Cutoff for total count. Counts below this cutoff value
will not be shown. Default is 0.}

\item{RowsCutoff}{Logical. If TRUE, filters based on cumulative percentage
cutoff. Default is FALSE.}

\item{digits}{Numeric. Number of decimal places for percentage columns.
Default is 2.}
}
\value{
Filtered data frame

A tibble with two columns:
\describe{
\item{\code{column}}{The unique elements from the specified column
(e.g., "DomArch").}
\item{\code{freq}}{The frequency of each element, i.e., the number of times
each element appears in the specified column.}
}
The tibble is filtered to only include elements that have a frequency
greater than or equal to \code{min.freq} and does not include elements with \code{NA}
values or those starting with a hyphen ("-").

A single string where elements are delimited by spaces. The function
performs necessary substitutions based on the \code{conversion_type} and cleans up
extraneous characters like newlines, tabs, and multiple spaces.

A tibble (tbl_df) with two columns:
\describe{
\item{\code{words}}{A column containing the individual words
(domains or domain architectures).}
\item{\code{freq}}{A column containing the frequency counts for each word.}
}

A tibble with the same structure as \code{x}, but filtered to include
only rows where the frequency is greater than or equal to \code{min.freq}.

A tibble summarizing the counts of occurrences of elements in
the \code{column}, grouped by the \code{by} column. The result includes the number
of occurrences (\code{count}) and is arranged in descending order of count.

A tibble summarizing the counts of unique domain architectures
(\code{DomArch}) per lineage (\code{Lineage}). The resulting table contains three
columns: \code{DomArch}, \code{Lineage}, and \code{count}, which indicates the frequency
of each domain architecture for each lineage. The results are arranged in
descending order of \code{count}.

A tibble summarizing each unique \code{DomArch}, along with the following
columns:
\itemize{
\item \code{totalcount}: The total occurrences of each \code{DomArch} across all lineages.
\item \code{totallin}: The total number of unique lineages in which each \code{DomArch}
appears.
The results are arranged in descending order of \code{totallin} and \code{totalcount}.
}

A tibble summarizing each unique combination of \code{GenContext},
\code{DomArch}, and \code{Lineage}, along with the following columns:
\itemize{
\item \code{GenContext}: The genomic context for each entry.
\item \code{DomArch}: The domain architecture for each entry.
\item \code{Lineage}: The lineage associated with each entry.
\item \code{count}: The total number of occurrences for each combination of
\code{GenContext}, \code{DomArch}, and \code{Lineage}.
}

The results are arranged in descending order of \code{count}.

Describe return, in detail

A tibble summarizing each unique combination of \code{GenContext} and
\code{Lineage}, along with the following columns:
\itemize{
\item \code{GenContext}: The genomic context for each entry.
\item \code{Lineage}: The lineage associated with each entry.
\item \code{count}: The total number of occurrences for each combination of
\code{GenContext} and \code{Lineage}.
}

The results are arranged in descending order of \code{count}.

A data frame with the following columns:
\itemize{
\item \code{{{ column }}}: Unique values from the specified column.
\item \code{totalcount}: The total count of occurrences for each unique value in
the specified column.
\item \code{IndividualCountPercent}: The percentage of each \code{totalcount} relative to
the overall count.
\item \code{CumulativePercent}: The cumulative percentage of total counts.
}
}
\description{
filter_by_doms filters a data frame by identifying exact domain matches
and either keeping or removing rows with the identified domain

Break string ELEMENTS into WORDS for domain architecture (DA) and genomic
context (GC)

Get word counts (wc) \link{DOMAINS (DA) or DOMAIN ARCHITECTURES (GC)}

Function to summarize and retrieve counts by Domains & Domains+Lineage

Function to retrieve counts of how many lineages a DomArch appears in

Creates a data frame with a totalcount column

This function is designed to sum the counts column by either Genomic Context or Domain Architecture and creates a totalcount column from those sums.
}
\note{
There is no need to make the domains 'regex safe', that will be handled by this function

Please refer to the source code if you have alternate file formats and/or
column names.
}
\examples{
\dontrun{
filter_by_doms()
}
\dontrun{
count_bycol(prot = my_data, column = "DomArch", min.freq = 10)
}
\dontrun{
tibble::tibble(DomArch = c("aaa+bbb", 
"a+b", "b+c", "b-c")) |> elements2words()
}

\dontrun{
tibble::tibble(DomArch = c("aaa+bbb", "a+b", "b+c", "b-c")) |>
    elements2words() |>
    words2wc()
}

\dontrun{
filter_freq()
}
\dontrun{
library(tidyverse)
tibble(DomArch = c("a+b", "a+b", "b+c", "a+b"), Lineage = c("l1", "l1", "l1", "l2")) |>
    summarize_bylin(query = "all")
}

\dontrun{
summ.DA.byLin()
}
\dontrun{
summ.DA()
}
\dontrun{
summ.GC.byDALin
}
\dontrun{
summ.GC.byLin()
}
\dontrun{
summ.GC()
}
\dontrun{
total_counts(pspa - gc_lin_counts, 0, "GC")
}
}
\author{
Samuel Chen, Janani Ravi
}
