% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summarize_bylin}
\alias{summarize_bylin}
\title{example: library(tidyverse); tibble(DomArch = c("a+b", "a+b", "b+c", "a+b"), Lineage = c("l1", "l1", "l1", "l2")) |>
summarize_bylin(query = "all")}
\usage{
summarize_bylin(prot = "prot", column = "DomArch", by = "Lineage", query)
}
\description{
example: library(tidyverse); tibble(DomArch = c("a+b", "a+b", "b+c", "a+b"), Lineage = c("l1", "l1", "l1", "l2")) |>
summarize_bylin(query = "all")
}
