% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summarizeByLineage}
\alias{summarizeByLineage}
\title{Summarize by Lineage}
\usage{
summarizeByLineage(prot = "prot", column = "DomArch", by = "Lineage", query)
}
\arguments{
\item{query}{}
}
\value{
Describe return, in detail
}
\description{
Summarize by Lineage
}
\examples{
\dontrun{
library(tidyverse)
tibble(DomArch = c("a+b", "a+b", "b+c", "a+b"), Lineage = c("l1", "l1", "l1", "l2")) |>
    summarizeByLineage(query = "all")
}

}
