% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summarizeDomArch}
\alias{summarizeDomArch}
\title{summarizeDomArch}
\usage{
summarizeDomArch(x)
}
\arguments{
\item{x}{}
}
\value{
Describe return, in detail
}
\description{
Function to retrieve counts of how many lineages a DomArch appears in
}
\examples{
\dontrun{
summarizeDomArch()
}
}
