% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summarizeDomArch_ByLineage}
\alias{summarizeDomArch_ByLineage}
\title{summarizeDomArch_ByLineage}
\usage{
summarizeDomArch_ByLineage(x)
}
\arguments{
\item{x}{}
}
\value{
Describe return, in detail
}
\description{
Function to summarize and retrieve counts by Domains & Domains+Lineage
}
\examples{
\dontrun{
summarizeDomArch_ByLineage()
}
}
