% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summarizeGenContext}
\alias{summarizeGenContext}
\title{summarizeGenContext}
\usage{
summarizeGenContext(x)
}
\arguments{
\item{x}{}
}
\value{
Describe return, in detail
}
\description{
summarizeGenContext
}
\examples{
\dontrun{
summarizeGenContext()
}
}
