% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summarizeGenContext_ByDomArchLineage}
\alias{summarizeGenContext_ByDomArchLineage}
\title{summarizeGenContext_ByDomArchLineage}
\usage{
summarizeGenContext_ByDomArchLineage(x)
}
\arguments{
\item{x}{}
}
\value{
Define return, in detail
}
\description{
summarizeGenContext_ByDomArchLineage
}
\examples{
\dontrun{
summarizeGenContext_ByDomArchLineage
}
}
