% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summarizeGenContext_ByLineage}
\alias{summarizeGenContext_ByLineage}
\title{summarizeGenContext_ByLineage}
\usage{
summarizeGenContext_ByLineage(x)
}
\arguments{
\item{x}{}
}
\value{
Describe return, in detail
}
\description{
summarizeGenContext_ByLineage
}
\examples{
\dontrun{
summarizeGenContext_ByLineage()
}
}
