% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipr2viz.R
\name{theme_genes2}
\alias{theme_genes2}
\title{Theme Genes2}
\usage{
theme_genes2()
}
\value{
A ggplot2 theme object.
}
\description{
Theme Genes2
}
\examples{
library(ggplot2)

# Create a sample plot using the custom theme
ggplot(mtcars, aes(x = wt, y = mpg)) +
    geom_point() +
    theme_genes2() +
    labs(title = "Car Weight vs MPG")

}
