% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre-msa-tree.R
\name{to_titlecase}
\alias{to_titlecase}
\alias{totitle,}
\alias{to_title}
\title{To Titlecase}
\usage{
to_titlecase(text, delimitter)
}
\arguments{
\item{x}{Character vector.}

\item{y}{Delimitter. Default is space (" ").}
}
\value{
A character vector in title case.
}
\description{
Translate string to Title Case w/ delimitter.
Changing case to 'Title Case'
}
\examples{
to_titlecase("hello world") 
to_titlecase("this is a test", "_") 
}
\seealso{
chartr, toupper, and tolower.
}
\author{
Andrie, Janani Ravi
}
