% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{total_counts}
\alias{total_counts}
\title{Total Counts}
\usage{
total_counts(prot, cutoff = 0, type = "GC")
}
\arguments{
\item{prot}{A data frame that must contain columns:
\itemize{\item Either 'GenContext.norep' or 'DomArch.norep' \item count}}

\item{cutoff}{Numeric. Cutoff for total count. Counts below cutoff value will not be shown. Default is 0.}

\item{type}{Character. Either "GC" for a total count by Genomic Context groupings or "DA" for a total count by Domain Architecture groupings.}
}
\description{
Creates a data frame with a totalcount column
}
\details{
This function is designed to sum the counts column by either Genomic Context or Domain Architecture and creates a totalcount column from those sums.
}
\note{
Please refer to the source code if you have alternate file formats and/or
column names.
}
\examples{
total_counts(pspa-gc_lin_counts,0,"GC")
}
