% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{totalgencontextordomarchcounts}
\alias{totalgencontextordomarchcounts}
\title{Total Counts}
\usage{
totalgencontextordomarchcounts(
  prot,
  column = "DomArch",
  lineage_col = "Lineage",
  cutoff = 90,
  RowsCutoff = FALSE,
  digits = 2
)
}
\arguments{
\item{prot}{A data frame that must contain columns:
\itemize{\item Either 'GenContext' or 'DomArch.norep' \item count}}

\item{column}{Character. The column to summarize}

\item{cutoff}{Numeric. Cutoff for total count. Counts below cutoff value will not be shown. Default is 0.}

\item{digits}{}
}
\value{
Define return, in detail
}
\description{
Creates a data frame with a totalcount column

This function is designed to sum the counts column by either Genomic Context or Domain Architecture and creates a totalcount column from those sums.
}
\note{
Please refer to the source code if you have alternate file formats and/or
column names.
}
\examples{
\dontrun{
totalgencontextordomarchcounts(pspa - gc_lin_counts, 0, "GC")
}
}
