% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{upset.plot}
\alias{upset.plot}
\title{UpSet Plot}
\usage{
upset.plot(query.sub = "toast_rack.sub", cutoff = 10,
  type = "da2doms")
}
\arguments{
\item{query.sub}{Data frame of protein homologs with the usual 11 columns +
additional word columns (0/1 format). Default is toast_rack.sub}

\item{cutoff}{Numeric. Cutoff for word frequency. Default is 10.}

\item{type}{Character. Either "da2doms" for Domains vs Domain Architectures
or "gc2da" for Domain Architectures (of neighbors) vs Genomic Contexts.
Default is "da2doms"}
}
\description{
UpSet plot for Domain Architectures vs Domains and
Genomic Contexts vs Domain Architectures.
}
\details{
For "da2doms" you would need the file DA.doms.wc as well as the
column query.sub$DomArch.norep

For "gc2da", you would need the file GC.DA.wc as well as the column
query.sub$GenContext.norep
}
\note{
Please refer to the source code if you have alternate file formats and/or
column names.
}
\examples{
upset.plot(pspa.sub, 10, "da2doms")
}
\author{
Janani Ravi
}
\keyword{Architectures,}
\keyword{Domain}
\keyword{Domains,}
\keyword{GenomicContexts}
\keyword{UpSetR,}
