% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotme.R
\name{validateCountDF}
\alias{validateCountDF}
\title{validateCountDF}
\usage{
validateCountDF(var)
}
\arguments{
\item{var}{A data frame whose columns are to be converted.}
}
\value{
A data frame with non-'n' columns converted to character type.
}
\description{
validateCountDF
}
\examples{
\dontrun{
new_df <- .all_non_n_cols_to_char(my_data)
}
}
