% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{wordcloud3}
\alias{wordcloud3}
\title{plotWordCloud3}
\usage{
wordcloud3(
  data,
  size = 1,
  minSize = 0,
  gridSize = 0,
  fontFamily = "Segoe UI",
  fontWeight = "bold",
  color = "random-dark",
  backgroundColor = "white",
  minRotation = -pi/4,
  maxRotation = pi/4,
  shuffle = TRUE,
  rotateRatio = 0.4,
  shape = "circle",
  ellipticity = 0.65,
  widgetsize = NULL,
  figPath = NULL,
  hoverFunction = NULL
)
}
\arguments{
\item{hoverFunction}{}
}
\description{
plotWordCloud3
}
