% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{wordcloud_element}
\alias{wordcloud_element}
\title{Wordclouds for the predominant domains, domain architectures.}
\usage{
wordcloud_element(
  query_data = "prot",
  colname = "DomArch",
  cutoff = 70,
  UsingRowsCutoff = FALSE
)
}
\arguments{
\item{query_data}{Data frame of protein homologs with the usual 11 columns +
additional word columns (0/1 format). Default is "prot".}

\item{type}{Character. Default is "da2doms" for Domain Architectures.
Other alternative: "gc2da" for Genomic Contexts.}
}
\description{
Wordclouds for the predominant domains (from DAs) and DAs (from GC)
}
\details{
For "da2doms" you would need the file DA.doms.wc as well as the
column query_data$DomArch.norep

For "gc2da", you would need the file GC.DA.wc as well as the column
query_data$GenContext.norep
}
\note{
Please refer to the source code if you have alternate file formats and/or
column names.
}
\examples{
wordcloud_element(prot, "da2doms", 10)
}
\author{
Janani Ravi
}
\keyword{Architectures,}
\keyword{Domain}
\keyword{Domains,}
\keyword{GenomicContexts}
