% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{words2WordCounts}
\alias{words2WordCounts}
\title{words2WordCounts}
\usage{
words2WordCounts(string)
}
\arguments{
\item{string}{A character string containing the elements (words) to count.
This would typically be a space-delimited string representing domain
architectures or genomic contexts.}
}
\value{
A tibble (tbl_df) with two columns:
\describe{
\item{\code{words}}{A column containing the individual words
(domains or domain architectures).}
\item{\code{freq}}{A column containing the frequency counts for each word.}
}
}
\description{
Get word counts (wc) \link{DOMAINS (DA) or DOMAIN ARCHITECTURES (GC)}
}
\examples{
\dontrun{
tibble::tibble(DomArch = c("aaa+bbb", "a+b", "b+c", "b-c")) |>
    elements2Words() |>
    words2WordCounts()
}

}
