% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHANGED-pre-msa-tree.R, R/pre-msa-tree.R
\name{write.MsaAAMultipleAlignment}
\alias{write.MsaAAMultipleAlignment}
\title{Write MsaAAMultpleAlignment Objects as algined fasta sequence}
\usage{
write.MsaAAMultipleAlignment(alignment, outpath)

write.MsaAAMultipleAlignment(alignment, outpath)
}
\arguments{
\item{alignment}{MsaAAMultipleAlignment object to be written as a fasta}

\item{outpath}{Where the resulting FASTA file should be written to}
}
\description{
MsaAAMultipleAlignment Objects are generated from calls to msaClustalOmega
and msaMuscle from the 'msa' package

Write MsaAAMultpleAlignment Objects as algined fasta sequence
MsaAAMultipleAlignment Objects are generated from calls to msaClustalOmega
and msaMuscle from the 'msa' package
}
\author{
Samuel Chen, Janani Ravi
}
