% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_job_queue.R
\name{write_proc_medians_table}
\alias{write_proc_medians_table}
\title{Write a table of 2 columns: 1) process and 2) median seconds}
\usage{
write_proc_medians_table(dir_job_results, filepath)
}
\arguments{
\item{dir_job_results}{\link{chr} path to MolEvolvR job_results}

\item{filepath}{path to save tsv file}
}
\value{
\link{tbl_df} 2 columns: 1) process and 2) median seconds

example: write_proc_medians_table(
"/data/scratch/janani/molevolvr_out/",
"/data/scratch/janani/molevolvr_out/log_tbl.tsv"
)
}
\description{
Write a table of 2 columns: 1) process and 2) median seconds
}
