% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_job_queue.R
\name{writeProcessRuntimeToYML}
\alias{writeProcessRuntimeToYML}
\title{Compute median process runtimes, then write a YAML list of the processes and
their median runtimes in seconds to the path specified by 'filepath'.}
\usage{
writeProcessRuntimeToYML(dir_job_results, filepath = NULL)
}
\arguments{
\item{dir_job_results}{\link{chr} path to MolEvolvR job_results directory}

\item{filepath}{\link{chr} path to save YAML file; if NULL,
uses ./molevol_scripts/log_data/job_proc_weights.yml}
}
\description{
The default value of filepath is the value of the env var
MOLEVOLVR_PROC_WEIGHTS, which writeProcessRuntimeToYML() also uses as its default
read location.
}
\examples{
\dontrun{
writeProcessRuntimeToYML(
    "/data/scratch/janani/molevolvr_out/",
    "/data/scratch/janani/molevolvr_out/log_tbl.yml"
)
}
}
