% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHANGED-pre-msa-tree.R, R/pre-msa-tree.R
\name{acc2fa}
\alias{acc2fa}
\title{acc2fa converts protein accession numbers to a fasta format.
Resulting fasta file is written to the outpath.}
\usage{
acc2fa(accessions, outpath, plan = "sequential")

acc2fa(accessions, outpath, plan = "sequential")
}
\arguments{
\item{accessions}{Character vector containing protein accession numbers to generate fasta sequences for.
Function may not work for vectors of length > 10,000}

\item{outpath}{\link{str}. Location where fasta file should be written to.}
}
\description{
acc2fa converts protein accession numbers to a fasta format.
Resulting fasta file is written to the outpath.

acc2fa converts protein accession numbers to a fasta format.
Resulting fasta file is written to the outpath.
}
\author{
Samuel Chen, Janani Ravi
}
\keyword{accnum,}
\keyword{fasta}
