% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acc2lin.R, R/lineage.R
\name{acc2lin}
\alias{acc2lin}
\title{acc2lin}
\usage{
acc2lin(
  accessions,
  assembly_path,
  lineagelookup_path,
  ipgout_path = NULL,
  plan = "multicore"
)

acc2lin(
  accessions,
  assembly_path,
  lineagelookup_path,
  ipgout_path = NULL,
  plan = "multicore"
)
}
\arguments{
\item{accessions}{Character vector of protein accessions}

\item{assembly_path}{String of the path to the assembly_summary path
This file can be generated using the "DownloadAssemblySummary()" function}

\item{lineagelookup_path}{String of the path to the lineage lookup file
(taxid to lineage mapping). This file can be generated using the}

\item{ipgout_path}{Path to write the results of the efetch run of the accessions
on the ipg database. If NULL, the file will not be written. Defaults to NULL}

\item{plan}{}
}
\value{
Describe return, in detail
}
\description{
This function combines 'efetch_ipg()' and 'ipg2lin()' to map a set
of protein accessions to their assembly (GCA_ID), tax ID, and lineage.

Function to map protein accession numbers to lineage

This function combines 'efetch_ipg()' and 'ipg2lin()' to map a set
of protein accessions to their assembly (GCA_ID), tax ID, and lineage.
}
\examples{
\dontrun{
acc2lin()
}
}
\author{
Samuel Chen, Janani Ravi
}
