% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acc2lin.R
\name{sinkReset}
\alias{sinkReset}
\alias{addlineage}
\alias{acc2Lineage}
\alias{efetchIPG}
\alias{IPG2Lineage}
\title{Sink Reset}
\usage{
sinkReset()

addlineage(
  df,
  acc_col = "AccNum",
  assembly_path,
  lineagelookup_path,
  ipgout_path = NULL,
  plan = "multicore"
)

acc2Lineage(
  accessions,
  assembly_path,
  lineagelookup_path,
  ipgout_path = NULL,
  plan = "multicore"
)

efetchIPG(accessions, out_path, plan = "multicore")

IPG2Lineage(
  accessions,
  ipg_file,
  refseq_assembly_path,
  genbank_assembly_path,
  lineagelookup_path
)
}
\arguments{
\item{assembly_path}{String of the path to the assembly_summary path
This file can be generated using the "DownloadAssemblySummary()" function}

\item{lineagelookup_path}{String of the path to the lineage lookup file
(taxid to lineage mapping). This file can be generated using the
"create_lineage_lookup()" function}

\item{ipgout_path}{Path to write the results of the efetch run of the accessions
on the ipg database. If NULL, the file will not be written. Defaults to NULL}

\item{plan}{}

\item{accessions}{Character vector of protein accessions}

\item{out_path}{Path to write the efetch results to}

\item{ipg_file}{Filepath to the file containing results of an efetch run on the
ipg database. The protein accession in 'accessions' should be contained in this
file}

\item{accnums}{Character vector containing the accession numbers to query on
the ipg database}
}
\value{
No return, but run to close all outstanding \code{sink()}s

Describe return, in detail

Describe return, in detail

Describe return, in detail

Describe return, in detail
}
\description{
This function combines 'efetchIPG()' and 'IPG2Lineage()' to map a set
of protein accessions to their assembly (GCA_ID), tax ID, and lineage.

Perform efetch on the ipg database and write the results to out_path

Takes the resulting file of an efetch run on the ipg database and
}
\examples{
\dontrun{
sinkReset()
}
\dontrun{
addlineage()
}
\dontrun{
acc2Lineage()
}
\dontrun{
efetchIPG()
}
\dontrun{
IPG2Lineage()
}

}
\author{
Samuel Chen, Janani Ravi
}
