% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acc2lin.R, R/lineage.R
\name{acc2Lineage}
\alias{acc2Lineage}
\title{acc2Lineage}
\usage{
acc2Lineage(
  accessions,
  assembly_path,
  lineagelookup_path,
  ipgout_path = NULL,
  plan = "multicore"
)

acc2Lineage(
  accessions,
  assembly_path,
  lineagelookup_path,
  ipgout_path = NULL,
  plan = "multicore"
)
}
\arguments{
\item{accessions}{Character vector of protein accessions}

\item{assembly_path}{String of the path to the assembly_summary path
This file can be generated using the "downloadAssemblySummary()" function}

\item{lineagelookup_path}{String of the path to the lineage lookup file
(taxid to lineage mapping). This file can be generated using the}

\item{ipgout_path}{Path to write the results of the efetch run of the accessions
on the ipg database. If NULL, the file will not be written. Defaults to NULL}

\item{plan}{}
}
\value{
A \code{data.table} that contains the lineage information, mapping protein
accessions to their tax IDs and lineages.
}
\description{
This function combines 'efetchIPG()' and 'IPG2Lineage()' to map a set
of protein accessions to their assembly (GCA_ID), tax ID, and lineage.

Function to map protein accession numbers to lineage

This function combines 'efetchIPG()' and 'IPG2Lineage()' to map a set
of protein accessions to their assembly (GCA_ID), tax ID, and lineage.
}
\examples{
\dontrun{
acc2Lineage()
}
}
\author{
Samuel Chen, Janani Ravi
}
