% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acc2lin.R, R/lineage.R
\name{add_lins}
\alias{add_lins}
\title{Add Lineages}
\usage{
add_lins(
  df,
  acc_col = "AccNum",
  assembly_path,
  lineagelookup_path,
  ipgout_path = NULL,
  plan = "multicore"
)

add_lins(
  df,
  acc_col = "AccNum",
  assembly_path,
  lineagelookup_path,
  ipgout_path = NULL,
  plan = "multicore"
)
}
\arguments{
\item{plan}{}
}
\value{
Describe return, in detail
}
\description{
Add Lineages

add_lins
}
\examples{
\dontrun{
add_lins()
}
}
