% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acc2lin.R
\name{add_lins}
\alias{add_lins}
\title{Add lineages}
\usage{
add_lins(
  df,
  acc_col = "AccNum",
  assembly_path,
  lineagelookup_path,
  ipgout_path = NULL,
  plan = "multicore"
)
}
\description{
Add lineages
}
