% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acc2lin.R, R/lineage.R
\name{addLineage}
\alias{addLineage}
\title{addLineage}
\usage{
addLineage(
  df,
  acc_col = "AccNum",
  assembly_path,
  lineagelookup_path,
  ipgout_path = NULL,
  plan = "multicore"
)

addLineage(
  df,
  acc_col = "AccNum",
  assembly_path,
  lineagelookup_path,
  ipgout_path = NULL,
  plan = "multicore"
)
}
\arguments{
\item{plan}{}
}
\value{
Describe return, in detail
}
\description{
addLineage

addLineage
}
\examples{
\dontrun{
addLineage()
}
}
