% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acc2lin.R, R/lineage.R
\name{addLineage}
\alias{addLineage}
\title{addLineage}
\usage{
addLineage(
  df,
  acc_col = "AccNum",
  assembly_path,
  lineagelookup_path,
  ipgout_path = NULL,
  plan = "multicore"
)

addLineage(
  df,
  acc_col = "AccNum",
  assembly_path,
  lineagelookup_path,
  ipgout_path = NULL,
  plan = "multicore"
)
}
\arguments{
\item{df}{A \code{data.frame} containing the input data. One column must contain
the accession numbers.}

\item{acc_col}{A string specifying the column name in \code{df} that holds the
accession numbers. Defaults to \code{"AccNum"}.}

\item{assembly_path}{A string specifying the path to the \code{assembly_summary.txt}
file. This file contains metadata about assemblies.}

\item{lineagelookup_path}{A string specifying the path to the lineage lookup
file, which contains a mapping from tax IDs to their corresponding lineages.}

\item{ipgout_path}{(Optional) A string specifying the path where IPG database
fetch results will be saved. If \code{NULL}, the results are not written to a file.}

\item{plan}{}
}
\value{
A \code{data.frame} that combines the original \code{df} with the lineage
information.
}
\description{
addLineage

addLineage
}
\examples{
\dontrun{
addLineage()
}
}
