% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineage.R
\name{addlineage}
\alias{addlineage}
\title{addlineage}
\usage{
addlineage(
  df,
  acc_col = "AccNum",
  assembly_path,
  lineagelookup_path,
  ipgout_path = NULL,
  plan = "multicore"
)
}
\arguments{
\item{plan}{}
}
\description{
addlineage
}
