% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanup.R
\name{cleanFAHeaders}
\alias{cleanFAHeaders}
\title{Cleanup FASTA Header}
\usage{
cleanFAHeaders(fasta)
}
\arguments{
\item{fasta}{An \link{XStringSet} object representing the sequences from a
FASTA file. The sequence names (headers) will be adjusted for uniqueness
and sanitized.}
}
\value{
\link{XStringSet} fasta with adjusted names (headers)
}
\description{
Parse accesion numbers from fasta and add a
suffix of the ith occurence to handle duplicates
}
\examples{
\dontrun{
AAStringSet(c("xxx" = "ATCG", "xxx" = "GGGC")) |>
    cleanFAHeaders()
}
}
