% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanup.R
\name{cleanLineage}
\alias{cleanLineage}
\title{Cleanup Lineage}
\usage{
cleanLineage(prot, lins_rename)
}
\arguments{
\item{lins_rename}{}
}
\value{
Describe return, in detail
}
\description{
Cleanup Lineage
}
\examples{
\dontrun{
cleanLineage()
}
}
