% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanup.R
\name{cleanup_GeneDesc}
\alias{cleanup_GeneDesc}
\title{Return trailing period that occurs in GeneDesc column}
\usage{
cleanup_GeneDesc(prot, column)
}
\description{
Return trailing period that occurs in GeneDesc column
}
