% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanup.R
\name{cleanup_species}
\alias{cleanup_species}
\title{Cleanup Species}
\usage{
cleanup_species(prot, remove_empty = FALSE)
}
\arguments{
\item{prot}{A data frame that contains columns 'Species' and 'GenContext.norep'}

\item{remove_empty}{Boolean. If TRUE, rows with empty/unnecessary values in 'GenContext.norep' are removed}
}
\description{
Cleans up the species column of a data frame by removing certain characters and rows.
}
\details{
This function removes unneccessary characters from the 'Species' column.
Certain rows may also removed from the table based on values in the 'GenContext.norep' column.
A cleaned up version of the data table is returned.
}
\examples{
cleanup_species(pspa.sub,TRUE)
}
