% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_analysis.R
\name{combineFullAnalysis}
\alias{combineFullAnalysis}
\title{Combining full_analysis files}
\usage{
combineFullAnalysis(inpath, ret = FALSE)
}
\arguments{
\item{ret}{}
}
\description{
Combining full_analysis files
}
