% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHANGED-pre-msa-tree.R, R/pre-msa-tree.R
\name{convert2TitleCase}
\alias{convert2TitleCase}
\alias{totitle,}
\alias{to_title}
\title{Changing case to 'Title Case'}
\usage{
convert2TitleCase(text, delimitter)

to_titlecase(text, delimitter)
}
\arguments{
\item{x}{Character vector.}

\item{y}{Delimitter. Default is space (" ").}
}
\value{
Character vector with the input strings converted to title case.

A character vector in title case.
}
\description{
Translate string to Title Case w/ delimitter.

Translate string to Title Case w/ delimitter.
Changing case to 'Title Case'
}
\examples{
# Convert a single string to title case
convert2TitleCase("hello world") # Returns "Hello World"

convert2TitleCase("hello world") 
convert2TitleCase("this is a test", "_") 
}
\seealso{
chartr, toupper, and tolower.

chartr, toupper, and tolower.
}
\author{
Andrie, Janani Ravi
}
