% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{convert_fa2tre}
\alias{convert_fa2tre}
\title{convert_fa2tre}
\usage{
convert_fa2tre(
  fa_path = here("data/alns/pspa_snf7.fa"),
  tre_path = here("data/alns/pspa_snf7.tre"),
  fasttree_path = here("src/FastTree")
)
}
\arguments{
\item{fasttree_path}{}
}
\description{
convert_fa2tre
}
