% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{count_bycol}
\alias{count_bycol}
\title{Count Bycol}
\usage{
count_bycol(prot = prot, column = "DomArch", min.freq = 1)
}
\arguments{
\item{min.freq}{}
}
\value{
Describe return, in detail
}
\description{
Count Bycol
}
\examples{
\dontrun{
count_bycol()
}
}
