% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{countbycolumn}
\alias{countbycolumn}
\title{Count Bycol}
\usage{
countbycolumn(prot = prot, column = "DomArch", min.freq = 1)
}
\arguments{
\item{min.freq}{}
}
\value{
Describe return, in detail
}
\description{
Count Bycol
}
\examples{
\dontrun{
countbycolumn()
}
}
