% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msa.R
\name{createMSA_Kalign}
\alias{createMSA_Kalign}
\title{Function to generate MSA using kalign}
\usage{
createMSA_Kalign(fa_file = "", outfile = "")
}
\arguments{
\item{outfile}{}
}
\description{
Function to generate MSA using kalign
}
