% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineage.R
\name{downloadAssemblySummary}
\alias{downloadAssemblySummary}
\title{Download the combined assembly summaries of genbank and refseq}
\usage{
downloadAssemblySummary(
  outpath,
  keep = c("assembly_accession", "taxid", "species_taxid", "organism_name")
)
}
\arguments{
\item{outpath}{String of path where the assembly summary file should be
written}

\item{keep}{Character vector containing which columns should be retained and
downloaded}
}
\value{
A tab-separated file containing the assembly summary. The function
does notreturn any value but writes the output directly to the specified file.
}
\description{
Download the combined assembly summaries of genbank and refseq
}
\examples{
\dontrun{
downloadAssemblySummary(outpath = "assembly_summary.tsv", 
     keep = c("assembly_accession", "taxid", "organism_name"))
}
}
\author{
Samuel Chen, Janani Ravi
}
