% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineage.R
\name{efetch_ipg}
\alias{efetch_ipg}
\title{efetch_ipg}
\usage{
efetch_ipg(accessions, out_path, plan = "multicore")
}
\arguments{
\item{accessions}{Character vector containing the accession numbers to query on
the ipg database}

\item{out_path}{Path to write the efetch results to}

\item{plan}{}
}
\description{
Perform efetch on the ipg database and write the results to out_path
}
\author{
Samuel Chen, Janani Ravi
}
