% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acc2lin.R, R/lineage.R
\name{efetchIPG}
\alias{efetchIPG}
\title{efetchIPG}
\usage{
efetchIPG(accessions, out_path, plan = "multicore")

efetchIPG(accessions, out_path, plan = "multicore")
}
\arguments{
\item{accessions}{Character vector containing the accession numbers to query on
the ipg database}

\item{out_path}{Path to write the efetch results to}

\item{plan}{}

\item{accnums}{Character vector containing the accession numbers to query on
the ipg database}
}
\value{
No return value. The function writes the fetched results to \code{out_path}.
}
\description{
Perform efetch on the ipg database
and write the results to out_path

Perform efetch on the ipg database and write the results to out_path
}
\examples{
\dontrun{
efetchIPG()
}
}
\author{
Samuel Chen, Janani Ravi
}
