% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{elements2words}
\alias{elements2words}
\title{Break string ELEMENTS into WORDS for domain architecture (DA) and genomic
context (GC)}
\usage{
elements2words(prot, column = "DomArch", conversion_type = "da2doms")
}
\value{
\link{string} with words delimited by spaces

example: tibble::tibble(DomArch = c("aaa+bbb", "a+b", "b+c", "b-c")) |> elements2words()
}
\description{
Break string ELEMENTS into WORDS for domain architecture (DA) and genomic
context (GC)
}
