% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{filter_freq}
\alias{filter_freq}
\title{Filter Frequency}
\usage{
filter_freq(x, min.freq)
}
\arguments{
\item{min.freq}{}
}
\value{
Describe return, in detail
}
\description{
Filter Frequency
}
\examples{
\dontrun{
filter_freq()
}
}
