% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R, R/utils.R
\name{filterByDomains}
\alias{filterByDomains}
\title{filterByDomains}
\usage{
filterByDomains(job_dir)

filterByDomains(job_dir)
}
\arguments{
\item{job_dir}{The directory containing the job files.}

\item{prot}{Dataframe to filter}

\item{column}{Column to search for domains in (DomArch column)}

\item{doms_keep}{Vector of domains that must be identified within column in order for
observation to be kept}

\item{doms_remove}{Vector of domains that, if found within an observation, will be removed}

\item{ignore.case}{Should the matching be non case sensitive}
}
\value{
Filtered data frame

None. Writes the filtered data to a file.
}
\description{
filterByDomains filters a data frame by identifying exact domain matches
and either keeping or removing rows with the identified domain

Filters the data by domains and writes the filtered data to a file.
}
\note{
There is no need to make the domains 'regex safe', that will be handled by this function
}
\examples{
\dontrun{
filterByDomains()
}
}
\author{
Samuel Chen, Janani Ravi
}
