% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{filterByFrequency}
\alias{filterByFrequency}
\title{filterByFrequency}
\usage{
filterByFrequency(x, min.freq)
}
\arguments{
\item{x}{A tibble (tbl_df) containing at least two columns: one for
elements (e.g., \code{words}) and one for their frequency (e.g., \code{freq}).}

\item{min.freq}{A numeric value specifying the minimum frequency threshold.
Only elements with frequencies greater than or equal to this value will be
retained.}
}
\value{
A tibble with the same structure as \code{x}, but filtered to include
only rows where the frequency is greater than or equal to \code{min.freq}.
}
\description{
Function to filter based on frequencies
}
\examples{
\dontrun{
filterByFrequency()
}
}
