% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{filterbyfrequency}
\alias{filterbyfrequency}
\title{Filter Frequency}
\usage{
filterbyfrequency(x, min.freq)
}
\arguments{
\item{min.freq}{}
}
\value{
Describe return, in detail
}
\description{
Filter Frequency
}
\examples{
\dontrun{
filterbyfrequency()
}
}
