% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHANGED-pre-msa-tree.R, R/pre-msa-tree.R
\name{get_accnums_from_fasta_file}
\alias{get_accnums_from_fasta_file}
\title{Get accnums from fasta file}
\usage{
get_accnums_from_fasta_file(fasta_file)

get_accnums_from_fasta_file(fasta_file)
}
\arguments{
\item{fasta_file}{}
}
\description{
Get accnums from fasta file

get_accnums_from_fasta_file
}
