% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHANGED-pre-msa-tree.R, R/pre-msa-tree.R
\name{getAccNumFromFA}
\alias{getAccNumFromFA}
\title{getAccNumFromFA}
\usage{
getAccNumFromFA(fasta_file)

getAccNumFromFA(fasta_file)
}
\arguments{
\item{fasta_file}{}
}
\description{
getAccNumFromFA

getAccNumFromFA
}
