% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accounting_db.R
\name{getCon}
\alias{getCon}
\title{Gets a connection to the accounting db. You should
close this connection later via dbDisconnect(con).}
\usage{
getCon()
}
\value{
A DBI connection object
}
\description{
Gets a connection to the accounting db. You should
close this connection later via dbDisconnect(con).
}
