% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineage.R
\name{ipg2lin}
\alias{ipg2lin}
\title{ipg2lin}
\usage{
ipg2lin(
  accessions,
  ipg_file,
  refseq_assembly_path,
  genbank_assembly_path,
  lineagelookup_path
)
}
\arguments{
\item{accessions}{Character vector of protein accessions}

\item{ipg_file}{Path to the file containing results of an efetch run on the
ipg database. The protein accession in 'accessions' should be contained in this
file}

\item{lineagelookup_path}{String of the path to the lineage lookup file
(taxid to lineage mapping). This file can be generated using the
"create_lineage_lookup()" function}
}
\description{
Takes the resulting file of an efetch run on the ipg database and
append lineage, and taxid columns
}
\author{
Samuel Chen, Janani Ravi
}
