% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acc2lin.R, R/lineage.R
\name{IPG2Lineage}
\alias{IPG2Lineage}
\title{IPG2Lineage}
\usage{
IPG2Lineage(
  accessions,
  ipg_file,
  refseq_assembly_path,
  genbank_assembly_path,
  lineagelookup_path
)

IPG2Lineage(
  accessions,
  ipg_file,
  refseq_assembly_path,
  genbank_assembly_path,
  lineagelookup_path
)
}
\arguments{
\item{accessions}{Character vector of protein accessions}

\item{ipg_file}{Path to the file containing results of an efetch run on the
ipg database. The protein accession in 'accessions' should be contained in this
file}

\item{lineagelookup_path}{String of the path to the lineage lookup file
(taxid to lineage mapping). This file can be generated using the
"createLineageLookup()" function}

\item{assembly_path}{String of the path to the assembly_summary path
This file can be generated using the \link[MolEvolvR]{downloadAssemblySummary} function}
}
\value{
A \code{data.table} with the lineage information for the provided protein
accessions.
}
\description{
Takes the resulting file
of an efetch run on the ipg database and

Takes the resulting file of an efetch run on the ipg database and
append lineage, and taxid columns
}
\examples{
\dontrun{
IPG2Lineage()
}

}
\author{
Samuel Chen, Janani Ravi
}
