% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotLineageDomainRepeats}
\alias{plotLineageDomainRepeats}
\title{Lineage Domain Repeats Plot}
\usage{
plotLineageDomainRepeats(query_data, colname, query_DAdoms)
}
\arguments{
\item{query_data}{Data frame containing protein homolog data, including
relevant domain architectures and lineages.}

\item{colname}{Character. The name of the column in query_data that contains
domain architectures or other structural information.}

\item{query_DAdoms}{List or data frame containing 'domains' element,
character vector of domain names. (NEW ARGUMENT)}
}
\value{
A ggplot object representing a heatmap (tile plot) of domain repeat
counts across different lineages, with color intensity representing the
occurrence of domains.
}
\description{
Lineage Domain Repeats Plot
}
\examples{
\dontrun{
plotLineageDomainRepeats()
}
}
