% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotme.R
\name{prepareSingleColumnParams}
\alias{prepareSingleColumnParams}
\title{prepareSingleColumnParams}
\usage{
prepareSingleColumnParams(df, col_num, root)
}
\arguments{
\item{df}{A data frame containing the data to be processed.}

\item{col_num}{An integer representing the column number to process.}

\item{root}{A string representing the root node for the treemap.}
}
\value{
A data frame containing parameters for the specified column for
treemap visualization.
}
\description{
prepareSingleColumnParams
}
\examples{
\dontrun{
df <- data.frame(Category = c("A", "A", "B", "B", "C"),
                 n = c(10, 20, 30, 40, 50))
params <- prepareSingleColumnParams(df, col_num = 1, root = "Root")
params
}
}
