% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineage.R
\name{proteinAcc2TaxID_old}
\alias{proteinAcc2TaxID_old}
\title{proteinAcc2TaxID_old}
\usage{
proteinAcc2TaxID_old(accessions, out_path, plan = "multicore")
}
\arguments{
\item{accessions}{Character vector containing the accession numbers to query on
the ipg database}

\item{out_path}{Path to write the efetch results to}

\item{plan}{}
}
\description{
Perform elink to go from protein database to taxonomy database
and write the resulting file of taxid and lineage to out_path
}
\author{
Samuel Chen, Janani Ravi
}
