% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanup.R
\name{remove_empty}
\alias{remove_empty}
\title{Remove empty rows by column}
\usage{
remove_empty(prot, by_column = "DomArch")
}
\arguments{
\item{prot}{A data frame containing 'DomArch', 'Species', 'GenContext', 'ClustName' columns.}

\item{by_column}{Column by which empty rows should be removed to domain+domain -> domain(s).
Default column is 'DomArch'. Can also take the following as input, 'Species', 'GenContext', 'ClustName'.}
}
\description{
Removes empty rows in the specified column.
}
\details{
This function ...
The original data frame is returned with the corresponding cleaned up column.
}
\examples{
remove_empty(prot, "DomArch")
}
