% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanup.R
\name{removeAsterisks}
\alias{removeAsterisks}
\title{Remove Astrk}
\usage{
removeAsterisks(query_data, colname = "GenContext")
}
\arguments{
\item{colname}{}
}
\value{
Describe return, in detail
}
\description{
Remove the asterisks from a column of data
Used for removing * from GenContext columns
}
\examples{
\dontrun{
removeAsterisks()
}
}
