% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanup.R
\name{selectLongestDuplicate}
\alias{selectLongestDuplicate}
\title{Pick Longer Duplicate}
\usage{
selectLongestDuplicate(prot, column)
}
\arguments{
\item{column}{}
}
\value{
Describe return, in detail
}
\description{
Pick Longer Duplicate
}
\examples{
\dontrun{
selectLongestDuplicate()
}
}
