% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{shortenLineage}
\alias{shortenLineage}
\title{shortenLineage}
\usage{
shortenLineage(data, colname = "Lineage", abr_len = 1)
}
\arguments{
\item{data}{A data frame that contains a column with lineage names to be
shortened.}

\item{colname}{Character. The name of the column in the data frame containing
the lineage strings to be shortened. Default is \code{"Lineage"}.}

\item{abr_len}{Integer. The number of characters to retain after the first
letter. If set to 1, only the first letter of each segment before the
delimiter (\code{>}) is retained. Default is 1.}
}
\value{
A modified data frame where the specified lineage column has been
shortened.
}
\description{
This function abbreviates lineage names by shortening the first part of the
string (up to a given delimiter).
}
\examples{
\dontrun{
df <- data.frame(Lineage = c("Bacteria>Firmicutes>Clostridia",
"Archaea>Euryarchaeota>Thermococci"))
shortened_df <- shortenLineage(df, colname = "Lineage", abr_len = 1)
shortened_df
}
}
