% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summ.DA.byLin}
\alias{summ.DA.byLin}
\title{summ.DA.byLin}
\usage{
summ.DA.byLin(x)
}
\arguments{
\item{x}{}
}
\value{
Describe return, in detail
}
\description{
Function to summarize and retrieve counts by Domains & Domains+Lineage
}
\examples{
\dontrun{
summ.DA.byLin()
}
}
