% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summ.DA}
\alias{summ.DA}
\title{summ.DA}
\usage{
summ.DA(x)
}
\arguments{
\item{x}{}
}
\value{
Describe return, in detail
}
\description{
Function to retrieve counts of how many lineages a DomArch appears in
}
\examples{
\dontrun{
summ.DA()
}
}
