% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summarize_bylin}
\alias{summarize_bylin}
\title{Summarize by Lineage}
\usage{
summarize_bylin(prot = "prot", column = "DomArch", by = "Lineage", query)
}
\arguments{
\item{prot}{A dataframe or tibble containing the data.}

\item{column}{A string representing the column to be summarized
(e.g., \code{DomArch}). Default is "DomArch".}

\item{by}{A string representing the grouping column (e.g., \code{Lineage}).
Default is "Lineage".}

\item{query}{A string specifying the query pattern for filtering the target
column. Use "all" to skip filtering and include all rows.}
}
\value{
A tibble summarizing the counts of occurrences of elements in
the \code{column}, grouped by the \code{by} column. The result includes the number
of occurrences (\code{count}) and is arranged in descending order of count.
}
\description{
Summarize by Lineage
}
\examples{
\dontrun{
library(tidyverse)
tibble(DomArch = c("a+b", "a+b", "b+c", "a+b"), Lineage = c("l1", "l1", "l1", "l2")) |>
    summarize_bylin(query = "all")
}

}
