% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_leaves.R
\name{to_titlecase}
\alias{to_titlecase}
\alias{totitle,}
\alias{to_title}
\title{Changing case to 'Title Case'}
\usage{
to_titlecase(text, delimitter)
}
\arguments{
\item{x}{Character vector.}

\item{y}{Delimitter. Default is space (" ").}
}
\description{
Translate string to Title Case w/ delimitter.
}
\seealso{
chartr, toupper, and tolower.
}
\author{
Andrie, Janani Ravi
}
