% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{total_counts}
\alias{total_counts}
\title{Total Counts}
\usage{
total_counts(
  prot,
  column = "DomArch",
  lineage_col = "Lineage",
  cutoff = 90,
  RowsCutoff = FALSE,
  digits = 2
)
}
\arguments{
\item{prot}{A data frame that must contain columns:
\itemize{\item Either 'GenContext' or 'DomArch.norep' \item count}}

\item{column}{Character. The column to summarize, default is "DomArch".}

\item{lineage_col}{Character. The name of the lineage column, default is
"Lineage".}

\item{cutoff}{Numeric. Cutoff for total count. Counts below this cutoff value
will not be shown. Default is 0.}

\item{RowsCutoff}{Logical. If TRUE, filters based on cumulative percentage
cutoff. Default is FALSE.}

\item{digits}{Numeric. Number of decimal places for percentage columns.
Default is 2.}
}
\value{
A data frame with the following columns:
\itemize{
\item \code{{{ column }}}: Unique values from the specified column.
\item \code{totalcount}: The total count of occurrences for each unique value in
the specified column.
\item \code{IndividualCountPercent}: The percentage of each \code{totalcount} relative to
the overall count.
\item \code{CumulativePercent}: The cumulative percentage of total counts.
}
}
\description{
Creates a data frame with a totalcount column

This function is designed to sum the counts column by either Genomic Context or Domain Architecture and creates a totalcount column from those sums.
}
\note{
Please refer to the source code if you have alternate file formats and/or
column names.
}
\examples{
\dontrun{
total_counts(pspa - gc_lin_counts, 0, "GC")
}
}
