% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{upset.plot}
\alias{upset.plot}
\title{UpSet Plot}
\usage{
upset.plot(
  query_data = "toast_rack.sub",
  colname = "DomArch",
  cutoff = 90,
  RowsCutoff = FALSE,
  text.scale = 1.5,
  point.size = 2.2,
  line.size = 0.8
)
}
\arguments{
\item{query_data}{Data frame of protein homologs with the usual 11 columns +
additional word columns (0/1 format). Default is toast_rack.sub}

\item{cutoff}{Numeric. Cutoff for word frequency. Default is 90.}

\item{text.scale}{Allows scaling of axis title, tick lables, and numbers above the intersection size bars.
text.scale can either take a universal scale in the form of an integer,
or a vector of specific scales in the format: c(intersection size title,
intersection size tick labels, set size title, set size tick labels, set names,
numbers above bars)}

\item{line.size}{}
}
\description{
UpSet plot for Domain Architectures vs Domains and
Genomic Contexts vs Domain Architectures.
}
\details{
For "da2doms" you would need the file DA.doms.wc as well as the
column query_data$DomArch.norep

For "gc2da", you would need the file GC.DA.wc as well as the column
query_data$GenContext.norep
}
\note{
Please refer to the source code if you have alternate file formats and/or
column names.
}
\examples{
\dontrun{
upset.plot(pspa.sub, 10, "da2doms")
}
}
\author{
Janani Ravi
}
\keyword{Architectures,}
\keyword{Domain}
\keyword{Domains,}
\keyword{GenomicContexts}
\keyword{UpSetR,}
