% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{words2wc}
\alias{words2wc}
\title{Get word counts (wc) \link{DOMAINS (DA) or DOMAIN ARCHITECTURES (GC)}}
\usage{
words2wc(string)
}
\value{
\link{tbl_df} table with 2 columns: 1) words & 2) counts/frequency
example: tibble::tibble(DomArch = c("aaa+bbb", "a+b", "b+c", "b-c")) |>
elements2words() |>
words2wc()
}
\description{
Get word counts (wc) \link{DOMAINS (DA) or DOMAIN ARCHITECTURES (GC)}
}
