% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre-msa-tree.R
\name{writeMSA_AA2FA}
\alias{writeMSA_AA2FA}
\title{writeMSA_AA2FA}
\usage{
writeMSA_AA2FA(alignment, outpath)
}
\arguments{
\item{alignment}{MsaAAMultipleAlignment object to be written as a fasta}

\item{outpath}{Where the resulting FASTA file should be written to}
}
\description{
Write MsaAAMultpleAlignment Objects as aligned fasta sequence
MsaAAMultipleAlignment Objects are generated from calls to msaClustalOmega
and msaMuscle from the 'msa' package
}
\author{
Samuel Chen, Janani Ravi
}
